refactor: centralize review tooling
