+++ mainappsrc/AutoML.py
@@ -242,18 +242,7 @@ from gui import messagebox, logger, add_treeview_scrollbars
 from gui.button_utils import enable_listbox_hover_highlight
 from gui.tooltip import ToolTip
 from gui.style_manager import StyleManager
-from gui.review_toolbox import (
-    ReviewToolbox,
-    ReviewData,
-    ReviewParticipant,
-    ReviewComment,
-    ParticipantDialog,
-    EmailConfigDialog,
-    ReviewScopeDialog,
-    UserSelectDialog as ReviewUserSelectDialog,
-    ReviewDocumentDialog,
-    VersionCompareDialog,
-)
+from gui.review_toolbox import ReviewData, ReviewParticipant, ReviewComment
 from functools import partial
 # Governance helper class
 from mainappsrc.governance_manager import GovernanceManager
@@ -2064,4451 +2053,7 @@ class AutoMLApp:
                 img = None
         finally:
             temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            def req_lines(reqs):
-                return "; ".join(
-                    self.format_requirement_with_trace(r) for r in reqs
-                )
-
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(
-                    old_data.get("description", ""), new_data.get("description", "")
-                )
-                rat_segments = [("Rationale: ", "black")] + diff_segments(
-                    old_data.get("rationale", ""), new_data.get("rationale", "")
-                )
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            req_segments = []
-
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            def req_lines(reqs):
-                return "; ".join(
-                    self.format_requirement_with_trace(r) for r in reqs
-                )
-
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(
-                    old_data.get("description", ""), new_data.get("description", "")
-                )
-                rat_segments = [("Rationale: ", "black")] + diff_segments(
-                    old_data.get("rationale", ""), new_data.get("rationale", "")
-                )
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            req_segments = []
-
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            def req_lines(reqs):
-                return "; ".join(
-                    self.format_requirement_with_trace(r) for r in reqs
-                )
-
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(
-                    old_data.get("description", ""), new_data.get("description", "")
-                )
-                rat_segments = [("Rationale: ", "black")] + diff_segments(
-                    old_data.get("rationale", ""), new_data.get("rationale", "")
-                )
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            req_segments = []
-
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            def req_lines(reqs):
-                return "; ".join(
-                    self.format_requirement_with_trace(r) for r in reqs
-                )
-
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(
-                    old_data.get("description", ""), new_data.get("description", "")
-                )
-                rat_segments = [("Rationale: ", "black")] + diff_segments(
-                    old_data.get("rationale", ""), new_data.get("rationale", "")
-                )
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            req_segments = []
-
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            def req_lines(reqs):
-                return "; ".join(
-                    f"[{r.get('id','')}] [{r.get('req_type','')}] {r.get('text','')}" for r in reqs
-                )
-
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(
-                    old_data.get("description", ""), new_data.get("description", "")
-                )
-                rat_segments = [("Rationale: ", "black")] + diff_segments(
-                    old_data.get("rationale", ""), new_data.get("rationale", "")
-                )
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            req_segments = []
-
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(old_data.get("description", ""), new_data.get("description", ""))
-                rat_segments = [("Rationale: ", "black")] + diff_segments(old_data.get("rationale", ""), new_data.get("rationale", ""))
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(old_data.get("description", ""), new_data.get("description", ""))
-                rat_segments = [("Rationale: ", "black")] + diff_segments(old_data.get("rationale", ""), new_data.get("rationale", ""))
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(old_data.get("description", ""), new_data.get("description", ""))
-                rat_segments = [("Rationale: ", "black")] + diff_segments(old_data.get("rationale", ""), new_data.get("rationale", ""))
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def capture_diff_diagram(self, top_event):
-        """Return an image of the FTA with diff colouring versus last version."""
-        if not self.versions:
-            return self.capture_page_diagram(top_event)
-
-        from io import BytesIO
-        from PIL import Image
-        import difflib
-
-        current = self.export_model_data(include_versions=False)
-        base_data = self.versions[-1]["data"]
-
-        def filter_events(data):
-            return [t for t in data.get("top_events", []) if t["unique_id"] == top_event.unique_id]
-
-        data1 = {"top_events": filter_events(base_data)}
-        data2 = {"top_events": filter_events(current)}
-
-        map1 = self.node_map_from_data(data1["top_events"])
-        map2 = self.node_map_from_data(data2["top_events"])
-
-        def build_conn_set(events):
-            conns = set()
-            def visit(d):
-                for ch in d.get("children", []):
-                    conns.add((d["unique_id"], ch["unique_id"]))
-                    visit(ch)
-            for t in events:
-                visit(t)
-            return conns
-
-        conns1 = build_conn_set(data1["top_events"])
-        conns2 = build_conn_set(data2["top_events"])
-
-        conn_status = {}
-        for c in conns1 | conns2:
-            if c in conns1 and c not in conns2:
-                conn_status[c] = "removed"
-            elif c in conns2 and c not in conns1:
-                conn_status[c] = "added"
-            else:
-                conn_status[c] = "existing"
-
-        status = {}
-        for nid in set(map1) | set(map2):
-            if nid in map1 and nid not in map2:
-                status[nid] = "removed"
-            elif nid in map2 and nid not in map1:
-                status[nid] = "added"
-            else:
-                if json.dumps(map1[nid], sort_keys=True) != json.dumps(map2[nid], sort_keys=True):
-                    status[nid] = "added"
-                else:
-                    status[nid] = "existing"
-
-        module = sys.modules.get(self.__class__.__module__)
-        FaultTreeNodeCls = getattr(module, 'FaultTreeNode', None)
-        if not FaultTreeNodeCls and self.top_events:
-            FaultTreeNodeCls = type(self.top_events[0])
-        new_roots = [FaultTreeNodeCls.from_dict(t) for t in data2["top_events"]]
-        removed_ids = [nid for nid, st in status.items() if st == "removed"]
-        for rid in removed_ids:
-            if rid in map1:
-                nd = map1[rid]
-                new_roots.append(FaultTreeNodeCls.from_dict(nd))
-
-        allow_ids = set()
-        def collect_ids(d):
-            allow_ids.add(d["unique_id"])
-            for ch in d.get("children", []):
-                collect_ids(ch)
-        if top_event.unique_id in map1:
-            collect_ids(map1[top_event.unique_id])
-        if top_event.unique_id in map2:
-            collect_ids(map2[top_event.unique_id])
-
-        node_objs = {}
-        def collect_nodes(n):
-            if n.unique_id not in node_objs:
-                node_objs[n.unique_id] = n
-            for ch in n.children:
-                collect_nodes(ch)
-        for r in new_roots:
-            collect_nodes(r)
-
-        def diff_segments(old, new):
-            matcher = difflib.SequenceMatcher(None, old, new)
-            segments = []
-            for tag, i1, i2, j1, j2 in matcher.get_opcodes():
-                if tag == "equal":
-                    segments.append((old[i1:i2], "black"))
-                elif tag == "delete":
-                    segments.append((old[i1:i2], "red"))
-                elif tag == "insert":
-                    segments.append((new[j1:j2], "blue"))
-                elif tag == "replace":
-                    segments.append((old[i1:i2], "red"))
-                    segments.append((new[j1:j2], "blue"))
-            return segments
-
-        def draw_segment_text(canvas, cx, cy, segments, font_obj):
-            lines = [[]]
-            for text, color in segments:
-                parts = text.split("\n")
-                for idx, part in enumerate(parts):
-                    if idx > 0:
-                        lines.append([])
-                    lines[-1].append((part, color))
-            line_height = font_obj.metrics("linespace")
-            total_height = line_height * len(lines)
-            start_y = cy - total_height / 2
-            for line in lines:
-                line_width = sum(font_obj.measure(part) for part, _ in line)
-                start_x = cx - line_width / 2
-                x = start_x
-                for part, color in line:
-                    if part:
-                        canvas.create_text(x, start_y, text=part, anchor="nw", fill=color, font=font_obj)
-                        x += font_obj.measure(part)
-                start_y += line_height
-
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-
-        def draw_connections(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_connections(ch)
-                return
-            region_width = 60
-            parent_bottom = (n.x, n.y + 20)
-            for i, ch in enumerate(n.children):
-                if ch.unique_id not in allow_ids:
-                    continue
-                parent_conn = (
-                    n.x - region_width / 2 + (i + 0.5) * (region_width / len(n.children)),
-                    parent_bottom[1],
-                )
-                child_top = (ch.x, ch.y - 25)
-                edge_st = conn_status.get((n.unique_id, ch.unique_id), "existing")
-                if status.get(n.unique_id) == "removed" or status.get(ch.unique_id) == "removed":
-                    edge_st = "removed"
-                color = "gray"
-                if edge_st == "added":
-                    color = "blue"
-                elif edge_st == "removed":
-                    color = "red"
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_conn, child_top, outline_color=color, line_width=1)
-                draw_connections(ch)
-
-        def draw_node(n):
-            if n.unique_id not in allow_ids:
-                for ch in n.children:
-                    draw_node(ch)
-                return
-            st = status.get(n.unique_id, "existing")
-            color = "dimgray"
-            if st == "added":
-                color = "blue"
-            elif st == "removed":
-                color = "red"
-
-            source = n if getattr(n, "is_primary_instance", True) else getattr(n, "original", n)
-            subtype_text = source.input_subtype if source.input_subtype else "N/A"
-            display_label = source.display_label
-            old_data = map1.get(n.unique_id)
-            new_data = map2.get(n.unique_id)
-            def req_lines(reqs):
-                return "; ".join(
-                    self.format_requirement_with_trace(r) for r in reqs
-                )
-
-            if old_data and new_data:
-                desc_segments = [("Desc: ", "black")] + diff_segments(
-                    old_data.get("description", ""), new_data.get("description", "")
-                )
-                rat_segments = [("Rationale: ", "black")] + diff_segments(
-                    old_data.get("rationale", ""), new_data.get("rationale", "")
-                )
-            else:
-                desc_segments = [("Desc: " + source.description, "black")]
-                rat_segments = [("Rationale: " + source.rationale, "black")]
-            req_segments = []
-
-            segments = [
-                (f"Type: {source.node_type}\n", "black"),
-                (f"Subtype: {subtype_text}\n", "black"),
-                (f"{display_label}\n", "black"),
-            ] + desc_segments + [("\n\n", "black")] + rat_segments
-
-            top_text = "".join(seg[0] for seg in segments)
-            bottom_text = n.name
-            fill = self.get_node_fill_color(n, getattr(canvas, "diagram_mode", None))
-            eff_x, eff_y = n.x, n.y
-            typ = n.node_type.upper()
-            items_before = canvas.find_all()
-            if typ in GATE_NODE_TYPES:
-                if n.gate_type and n.gate_type.upper() == "OR":
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_or_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-                else:
-                    if self.fta_drawing_helper:
-                        self.fta_drawing_helper.draw_rotated_and_gate_shape(canvas, eff_x, eff_y, scale=40, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-            else:
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_circle_event_shape(canvas, eff_x, eff_y, 45, top_text=top_text, bottom_text=bottom_text, fill=fill, outline_color=color, line_width=2)
-
-            items_after = canvas.find_all()
-            text_id = None
-            for item in items_after:
-                if item in items_before:
-                    continue
-                if canvas.type(item) == "text" and canvas.itemcget(item, "text") == top_text:
-                    text_id = item
-                    break
-
-            if text_id and any(c in ("red", "blue") for _, c in segments):
-                bbox = canvas.bbox(text_id)
-                cx = (bbox[0] + bbox[2]) / 2
-                cy = (bbox[1] + bbox[3]) / 2
-                canvas.itemconfigure(text_id, state="hidden")
-                draw_segment_text(canvas, cx, cy, segments, self.diagram_font)
-            for ch in n.children:
-                draw_node(ch)
-
-        for r in new_roots:
-            draw_connections(r)
-            draw_node(r)
-
-        existing_pairs = set()
-        for p in node_objs.values():
-            for ch in p.children:
-                existing_pairs.add((p.unique_id, ch.unique_id))
-        for (pid, cid), st in conn_status.items():
-            if st != "removed":
-                continue
-            if (pid, cid) in existing_pairs:
-                continue
-            if pid in node_objs and cid in node_objs and pid in allow_ids and cid in allow_ids:
-                parent = node_objs[pid]
-                child = node_objs[cid]
-                parent_pt = (parent.x, parent.y + 20)
-                child_pt = (child.x, child.y - 25)
-                if self.fta_drawing_helper:
-                    self.fta_drawing_helper.draw_90_connection(canvas, parent_pt, child_pt, outline_color="red", line_width=1)
-
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        ps = canvas.postscript(colormode="color", x=x, y=y, width=x2 - x, height=y2 - y)
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def metric_to_text(self, metric_type, value):
-        if value is None:
-            return "unknown"
-        disc = AutoML_Helper.discretize_level(value)
-        if metric_type == "maturity":
-            return "high maturity" if disc == 5 else "low maturity" if disc == 1 else f"a maturity of {disc}"
-        elif metric_type == "rigor":
-            return "high rigor" if disc == 5 else "low rigor" if disc == 1 else f"a rigor of {disc}"
-        elif metric_type == "severity":
-            return "high severity" if disc >= 3 else "low severity" if disc == 1 else f"a severity of {disc}"
-        else:
-            return str(disc)
-
-    def assurance_level_text(self, level):
-        mapping = {1:"PAL1", 2:"PAL2", 3:"PAL3", 4:"PAL4", 5:"PAL5"}
-        return mapping.get(level, str(level))
-
-    def calculate_cut_sets(self, node):
-        if not node.children:
-            return [{node.unique_id}]
-        gate = (node.gate_type or "AND").upper() if node.node_type.upper() in GATE_NODE_TYPES else "AND"
-        child_cut_sets = [self.calculate_cut_sets(child) for child in node.children]
-        if gate == "OR":
-            result = []
-            for cuts in child_cut_sets:
-                result.extend(cuts)
-            return result
-        elif gate == "AND":
-            result = [set()]
-            for cuts in child_cut_sets:
-                temp = []
-                for partial in result:
-                    for cs in cuts:
-                        temp.append(partial.union(cs))
-                result = temp
-            return result
-        else:
-            result = []
-            for cuts in child_cut_sets:
-                result.extend(cuts)
-            return result
-
-    def build_hierarchical_argumentation(self, node, indent=0):
-        indent_str = "    " * indent
-        node_name = node.user_name if node.user_name else f"Node {node.unique_id}"
-        details = f"{node.node_type}"
-        if node.input_subtype:
-            details += f" ({node.input_subtype})"
-        if node.description:
-            details += f": {node.description}"
-        metric_type = "maturity" if node.node_type.upper() in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"] else "rigor"
-        metric_descr = self.metric_to_text(metric_type, node.quant_value)
-        line = f"{indent_str}- {node_name} ({details}) -> {metric_descr}"
-        if node.rationale and node.node_type.upper() not in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"]:
-            line += f" [Rationale: {node.rationale.strip()}]"
-        child_lines = [self.build_hierarchical_argumentation(child, indent+1) for child in node.children]
-        if child_lines:
-            line += "\n" + "\n".join(child_lines)
-        return line
-
-    def build_hierarchical_argumentation_common(self, node, indent=0, described=None):
-        if described is None:
-            described = set()
-        indent_str = "    " * indent
-        node_name = node.user_name if node.user_name else f"Node {node.unique_id}"
-        if node.unique_id not in described:
-            details = f"{node.node_type}"
-            if node.input_subtype:
-                details += f" ({node.input_subtype})"
-            if node.description:
-                details += f": {node.description}"
-            described.add(node.unique_id)
-        else:
-            details = f"{node.node_type} (see common cause: {node_name})"
-        metric_type = "maturity" if node.node_type.upper() in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"] else "rigor"
-        metric_descr = self.metric_to_text(metric_type, node.quant_value)
-        line = f"{indent_str}- {node_name} ({details}) -> {metric_descr}"
-        if node.rationale and node.node_type.upper() not in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"]:
-            line += f" [Rationale: {node.rationale.strip()}]"
-        child_lines = [self.build_hierarchical_argumentation_common(child, indent+1, described) for child in node.children]
-        if child_lines:
-            line += "\n" + "\n".join(child_lines)
-        return line
-
-    def build_page_argumentation(self, page_node):
-        return self.build_hierarchical_argumentation(page_node)
-
-    def build_unified_recommendation_table(self):
-        """
-        Collect ALL nodes across ALL top-level events, group them by the
-        recommendation(s) they trigger, and return a single LongTable.
-        
-        *Only primary nodes (originals) are used so that clones are not duplicated.
-        Each node gets its own row, so large text can split across pages.
-        """
-        from reportlab.platypus import LongTable, Paragraph
-        from reportlab.lib import colors
-        from reportlab.lib.styles import ParagraphStyle, getSampleStyleSheet
-        from reportlab.platypus import TableStyle
-
-        style_sheet = getSampleStyleSheet()
-        body_style = style_sheet["BodyText"]
-        header_style = ParagraphStyle(
-            name="RecHeader",
-            parent=body_style,
-            fontSize=5,
-            leading=6,
-            wordWrap='CJK',
-            alignment=1
-        )
-
-        # 1) Gather ALL nodes from ALL top events.
-        # Assumes get_all_nodes_in_model() is defined to merge all nodes.
-        all_nodes = self.get_all_nodes_in_model()
-        if not all_nodes:
-            print("Debug: No nodes found in the entire model.")
-            return None
-
-        # 2) Filter out clones: only include primary instances.
-        primary_nodes = [n for n in all_nodes if n.is_primary_instance]
-
-        # 3) Build a mapping: recommendation text -> list of nodes that trigger it.
-        rec_to_nodes = {}
-        for node in primary_nodes:
-            # Only consider nodes with a quant_value and a nonempty description.
-            if node.quant_value is not None and node.description:
-                discrete = AutoML_Helper.discretize_level(node.quant_value)
-                extra_dict = dynamic_recommendations.get(discrete, {}).get("Extra Recommendations", {})
-                desc_lower = node.description.lower()
-                for keyword, rec_text in extra_dict.items():
-                    if keyword.lower() in desc_lower:
-                        rec_to_nodes.setdefault(rec_text, []).append(node)
-
-        if not rec_to_nodes:
-            print("Debug: No matching recommendations found for any node.")
-            return None
-
-        # 4) Build the table rows.
-        # We use two columns: "Extra Recommendation" and "Metric Details"
-        # For each recommendation, the first row shows the recommendation text and the details
-        # for the first node; subsequent rows leave the recommendation column blank.
-        data = [[
-            Paragraph("<b>Extra Recommendation</b>", header_style),
-            Paragraph("<b>Metric Details</b>", header_style)
-        ]]
-
-        for rec_text, nodes in rec_to_nodes.items():
-            first_row = True
-            for node in nodes:
-                # Use the node's display_label if it does not fall back to "Node ..."; otherwise, use the quant_value.
-                metric_str = (node.display_label 
-                              if node.display_label and not node.display_label.startswith("Node")
-                              else (f"{node.quant_value:.2f}" if node.quant_value is not None else "N/A"))
-                desc = (node.description or "N/A").strip().replace("\n", "<br/>")
-                rat = (node.rationale or "N/A").strip().replace("\n", "<br/>")
-                node_details = (
-                    f"{node.unique_id}: {node.name}"
-                    f"<br/><b>Metric:</b> {metric_str}"
-                    f"<br/><b>Desc:</b> {desc}"
-                    f"<br/><b>Rationale:</b> {rat}"
-                )
-                if first_row:
-                    data.append([
-                        Paragraph(rec_text, body_style),
-                        Paragraph(node_details, body_style)
-                    ])
-                    first_row = False
-                else:
-                    data.append([
-                        "",
-                        Paragraph(node_details, body_style)
-                    ])
-
-        # 5) Create and style the LongTable.
-        col_widths = [200, 450]
-        table = LongTable(data, colWidths=col_widths, repeatRows=1, splitByRow=True)
-        table.setStyle(TableStyle([
-            ('BACKGROUND', (0,0), (-1,0), colors.orange),
-            ('TEXTCOLOR', (0,0), (-1,0), colors.white),
-            ('GRID', (0,0), (-1,-1), 0.25, colors.grey),
-            ('VALIGN', (0,0), (-1,-1), 'TOP'),
-            ('FONTSIZE', (0,0), (-1,-1), 5),
-            ('LEFTPADDING', (0,0), (-1,-1), 2),
-            ('RIGHTPADDING', (0,0), (-1,-1), 2),
-            ('TOPPADDING', (0,0), (-1,-1), 2),
-            ('BOTTOMPADDING', (0,0), (-1,-1), 2),
-        ]))
-        return table
-
-    def build_base_events_table_html(self,root_node):
-        """
-        Traverse the fault tree starting from root_node and collect all base events (leaf nodes).
-        Return an HTML string representing a table with columns:
-        Node ID, Name, Score, Description, and Rationale.
-        
-        The table header cells are styled with an orange background (#FFCC99),
-        and multiline descriptions/rationales are preserved by converting newlines to <br>.
-        """
-        base_events = []
-        
-        # Collect all leaf nodes (base events).
-        def traverse(n):
-            if not n.children:
-                base_events.append(n)
-            else:
-                for child in n.children:
-                    traverse(child)
-        traverse(root_node)
-        
-        # Build the HTML table.
-        html_lines = []
-        html_lines.append('<table style="border-collapse: collapse; width: 100%;">')
-        
-        # Table header row with orange background.
-        html_lines.append(
-            '  <thead>'
-            '    <tr style="background-color: #FFCC99;">'
-            '      <th style="border: 1px solid #ccc; padding: 8px;">Node ID</th>'
-            '      <th style="border: 1px solid #ccc; padding: 8px;">Name</th>'
-            '      <th style="border: 1px solid #ccc; padding: 8px;">Score</th>'
-            '      <th style="border: 1px solid #ccc; padding: 8px;">Description</th>'
-            '      <th style="border: 1px solid #ccc; padding: 8px;">Rationale</th>'
-            '    </tr>'
-            '  </thead>'
-        )
-        html_lines.append('  <tbody>')
-        
-        for node in base_events:
-            node_id = str(node.unique_id)
-            name = node.name or f"Node {node.unique_id}"
-            score = f"{node.quant_value:.2f}" if node.quant_value is not None else "N/A"
-            
-            # Convert newlines to <br> for multiline display.
-            desc = node.description.strip().replace('\n', '<br>') if node.description else "N/A"
-            rat = node.rationale.strip().replace('\n', '<br>') if node.rationale else "N/A"
-            
-            row_html = (
-                '    <tr>'
-                f'      <td style="border: 1px solid #ccc; padding: 8px; vertical-align: top;">{node_id}</td>'
-                f'      <td style="border: 1px solid #ccc; padding: 8px; vertical-align: top;">{name}</td>'
-                f'      <td style="border: 1px solid #ccc; padding: 8px; vertical-align: top;">{score}</td>'
-                f'      <td style="border: 1px solid #ccc; padding: 8px; vertical-align: top;">{desc}</td>'
-                f'      <td style="border: 1px solid #ccc; padding: 8px; vertical-align: top;">{rat}</td>'
-                '    </tr>'
-            )
-            html_lines.append(row_html)
-        
-        html_lines.append('  </tbody>')
-        html_lines.append('</table>')
-        
-        return "\n".join(html_lines)
-
-    def build_argumentation(self, node):
-        if not node.children:
-            return ""
-        header = ""
-        if node.node_type.upper() == "TOP EVENT":
-            disc = AutoML_Helper.discretize_level(node.quant_value)
-            assurance_descr = self.assurance_level_text(disc)
-            severity_str = f"{node.severity}" if node.severity is not None else "N/A"
-            controllability_str = f"{node.controllability}" if node.controllability is not None else "N/A"
-            header += (
-                f"Prototype Assurance Level (PAL) Explanation:<br/>"
-                f"Based on the aggregated scores of its child nodes, this top event has been assigned an Prototype Assurance Level (PAL) of <b>{assurance_descr}</b> "
-                f"with a severity rating of <b>{severity_str}</b> and controllability <b>{controllability_str}</b>.<br/><br/>"
-            )
-            # Append the dynamically generated recommendations.
-            header += self.generate_recommendations_for_top_event(node) + "<br/>"
-        # Now generate the cut-set table.
-        nodes_by_id = {}
-        def map_nodes(n):
-            nodes_by_id[n.unique_id] = n
-            for child in n.children:
-                map_nodes(child)
-        map_nodes(node)
-        cut_sets = self.calculate_cut_sets(node)
-        cut_set_table = "Cut Set Table:<br/>"
-        for i, cs in enumerate(cut_sets, start=1):
-            cs_ids = ", ".join(f"Node {uid}" for uid in sorted(cs))
-            cut_set_table += f"Cut Set {i}: {cs_ids}<br/>"
-        node_definitions = "Node Definitions:<br/>"
-        unique_ids = set()
-        for cs in cut_sets:
-            unique_ids.update(cs)
-        for uid in sorted(unique_ids):
-            n = nodes_by_id.get(uid)
-            if n is None:
-                continue
-            subtype = n.input_subtype if n.input_subtype is not None else (
-                VALID_SUBTYPES["Confidence"][0] if n.node_type.upper() == "CONFIDENCE LEVEL"
-                else VALID_SUBTYPES.get("Prototype Assurance Level (PAL)", ["Default"])[0]
-            )
-            desc = n.description.strip() if n.description else "No description provided."
-            node_definitions += f"Node {uid}: {n.name}<br/>"
-            node_definitions += f"Type: {n.node_type}, Subtype: {subtype}<br/>"
-            node_definitions += f"Description: {desc}<br/><br/>"
-        diagram_note = "Cause-and-Effect Diagram is generated below.<br/>"
-        return header + cut_set_table + "<br/>" + node_definitions + "<br/>" + diagram_note
-       
-    def auto_create_argumentation(self, node, suppress_top_event_recommendations=False):
-        """
-        Generate qualitative argumentation text for a given node.
-        For a TOP EVENT (unless suppressed), include dynamic recommendations from the dictionary
-        filtered by the node’s description. For non–top-level nodes, simply display the node's input score,
-        description, and rationale.
-        """
-        level = AutoML_Helper.discretize_level(node.quant_value) if node.quant_value is not None else 1
-
-        if node.node_type.upper() == "TOP EVENT" and not suppress_top_event_recommendations:
-            assurance_descr = self.assurance_level_text(level)
-            severity_str = f"{node.severity}" if node.severity is not None else "N/A"
-            controllability_str = f"{node.controllability}" if node.controllability is not None else "N/A"
-            header = (
-                f"Prototype Assurance Level (PAL) Explanation:\n"
-                f"This top event is assigned an Prototype Assurance Level (PAL) of '{assurance_descr}' with a severity rating of {severity_str} and controllability {controllability_str}.\n\n"
-            )
-            # Instead of showing all dynamic recommendations, select only those triggered by the description.
-            rec_from_desc = self.get_recommendation_from_description(node.description, level)
-            if rec_from_desc:
-                base_arg = header + "Dynamic Recommendation:\n" + rec_from_desc
-            else:
-                # If no keyword found, show the full recommendations.
-                rec = dynamic_recommendations.get(level, {})
-                rec_lines = []
-                for category in ["Testing Requirements", "IFTD Responsibilities", "Preventive Maintenance Actions", "Relevant AVSC Guidelines"]:
-                    if category in rec:
-                        rec_lines.append(f"{category}: {rec[category]}")
-                rec_text = "\n".join(rec_lines)
-                base_arg = header + "Recommendations:\n" + rec_text
-        elif node.node_type.upper() == "TOP EVENT" and suppress_top_event_recommendations:
-            base_arg = f"Top Event: Input score: {node.quant_value:.2f}" if node.quant_value is not None else "Top Event: No input score provided."
-        else:
-            base_arg = f"Input score: {node.quant_value:.2f}" if node.quant_value is not None else "No input score provided."
-
-        own_text = ""
-        if node.description:
-            own_text += f"Description: {node.description}\n"
-        if node.rationale:
-            own_text += f"Rationale: {node.rationale}\n"
-        if not own_text:
-            own_text = "No additional details provided."
-            
-        return base_arg + "\n\n" + own_text
-
-    def generate_argumentation_report(self, event):
-        """
-        Generate dynamic assurance-level argumentation for a top-level event.
-        In this version, the event’s description is added at the very beginning,
-        followed by the Prototype Assurance Level (PAL) explanation (including the rationale behind its severity)
-        and the dynamic recommendations.
-        """
-        # Ensure a quant_value exists; default to 1.
-        quant = event.quant_value if event.quant_value is not None else 1
-        level = AutoML_Helper.discretize_level(quant)
-        assurance_level = self.assurance_level_text(level)
-        severity = event.severity if event.severity is not None else "N/A"
-        controllability = event.controllability if event.controllability is not None else "N/A"
-        
-        # Get dynamic recommendations from the dictionary.
-        rec = dynamic_recommendations.get(level, {})
-        test_req = rec.get("Testing Requirements", "N/A")
-        iftd_resp = rec.get("IFTD Responsibilities", "N/A")
-        maint_act = rec.get("Preventive Maintenance Actions", "N/A")
-        avsc_guid = rec.get("Relevant AVSC Guidelines", "N/A")
-        
-        # Get and clean up the top event’s description and rationale.
-        top_description = event.description.strip() if event.description and event.description.strip() else "N/A"
-        top_rationale = event.rationale.strip() if event.rationale and event.rationale.strip() else "N/A"
-        
-        text = (
-            f"Description:<br/>{top_description}<br/><br/>"
-            f"Prototype Assurance Level (PAL) Explanation:<br/>"
-            f"This top event is assigned an Prototype Assurance Level (PAL) of <b>{assurance_level}</b> "
-            f"with a severity rating of <b>{severity}</b> and controllability <b>{controllability}</b>.<br/>"
-            f"Rationale for Severity: {top_rationale}<br/><br/>"
-            #"Dynamic Recommendations:<br/>"
-            #f"<b>Testing Requirements:</b> {test_req}<br/>"
-            #f"<b>IFTD Responsibilities:</b> {iftd_resp}<br/>"
-            #f"<b>Preventive Maintenance Actions:</b> {maint_act}<br/>"
-            #f"<b>Relevant AVSC Guidelines:</b> {avsc_guid}<br/>"
-        )
-        return text
-
-    def get_extra_recommendations_list(self, description, level):
-        """
-        Given a node's description and its Prototype Assurance Level (PAL), return a list of extra recommendations.
-        This function iterates over all keys in the level's "Extra Recommendations" dictionary and
-        collects the recommendation text for every keyword found in the description.
-        """
-        if not description:
-            return []
-        desc = description.lower()
-        level_extras = dynamic_recommendations.get(level, {}).get("Extra Recommendations", {})
-        rec_list = []
-        for keyword, rec in level_extras.items():
-            if keyword.lower() in desc:
-                rec_list.append(rec)
-        return rec_list
-
-    def get_extra_recommendations_from_level(self,description, level):
-        """
-        Given a node's description and its Prototype Assurance Level (PAL) (1-5), look up keywords from the level's 
-        "Extra Recommendations" in the dynamic_recommendations dictionary. If any keyword is found in the description
-        (within a proximity of malfunction words), return the extra recommendations.
-        """
-        if not description:
-            return ""
-        desc = description.lower()
-        level_extras = dynamic_recommendations.get(level, {}).get("Extra Recommendations", {})
-        malfunction_words = ["unintended", "no", "not", "excessive", "incorrect"]
-        
-        recommendations = []
-        for keyword, rec in level_extras.items():
-            # Check if the keyword is present.
-            if re.search(r'\b' + re.escape(keyword) + r'\b', desc):
-                # Look for a malfunction word within 5 words of the keyword.
-                pattern = r'\b' + re.escape(keyword) + r'\b(?:\W+\w+){0,5}\W+(?:' + "|".join(malfunction_words) + r')\b'
-                if re.search(pattern, desc):
-                    recommendations.append(rec)
-        if recommendations:
-            return "\nExtra Testing Recommendations:\n" + "\n".join(f"- {r}" for r in recommendations)
-        return ""
-
-    def get_recommendation_from_description(self, description, level):
-        """
-        Given a node's description and its Prototype Assurance Level (PAL), this function iterates over all keys 
-        in the corresponding level's "Extra Recommendations" dictionary. It checks if each keyword 
-        appears in the description (in a case-insensitive manner) and concatenates all matching recommendations.
-        """
-        if not description:
-            return ""
-        desc = description.lower()
-        level_extras = dynamic_recommendations.get(level, {}).get("Extra Recommendations", {})
-        rec_list = []
-        for keyword, rec in level_extras.items():
-            if keyword.lower() in desc:
-                rec_list.append(rec)
-        return " ".join(rec_list)
-    
-    def analyze_common_causes(self, node):
-        occurrence = {}
-        def traverse(n):
-            if n.unique_id in occurrence:
-                occurrence[n.unique_id]["count"] += 1
-            else:
-                occurrence[n.unique_id] = {"node": n, "count": 1}
-            for child in n.children:
-                traverse(child)
-        traverse(node)
-        report_lines = ["Common Cause Analysis:"]
-        for uid, info in occurrence.items():
-            if info["count"] > 1:
-                n = info["node"]
-                name = n.user_name if n.user_name else f"Node {n.unique_id}"
-                report_lines.append(f" - {name} (Type: {n.node_type}) appears {info['count']} times. Description: {n.description or 'No description'}")
-        if len(report_lines) == 1:
-            report_lines.append(" None found.")
-        return "\n".join(report_lines)
-
-    def build_text_report(self, node, indent=0):
-        report = "    " * indent + f"{node.name} ({node.node_type}"
-        if node.node_type.upper() in GATE_NODE_TYPES:
-            report += f", {node.gate_type}"
-        report += ")"
-        if node.display_label:
-            report += f" => {node.display_label}"
-        arg_text = self.build_argumentation(node)
-        if arg_text:
-            report += f"\n{'    ' * (indent+1)}Argumentation: {arg_text}"
-        report += "\n\n"
-        for child in node.children:
-            report += self.build_text_report(child, indent+1)
-        return report
-
-    def all_children_are_base_events(self,node):
-        """
-        Return True if *every* child of 'node' is a base event 
-        (i.e. Confidence Level or Robustness Score).
-        """
-        # If node has no children, we treat it as "False" (it’s effectively a leaf, not a gate).
-        if not node.children:
-            return False
-
-        for child in node.children:
-            t = child.node_type.upper()
-            if t not in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"]:
-                return False
-        return True
-
-    def build_simplified_fta_model(self, top_event):
-        """
-        Build a simplified FTA model from the fault tree by including only the gate-level nodes.
-        If a node is a GATE, RIGOR LEVEL, or TOP EVENT but all its children are base events,
-        we will skip showing its gate_type.
-        """
-        nodes = []
-        edges = []
-
-        visited = set()
-
-        def traverse(node):
-            if node.unique_id in visited:
-                return
-            visited.add(node.unique_id)
-
-            node_type_up = node.node_type.upper()
-            node_info = {
-                "id": str(node.unique_id),
-                "label": node.name,
-            }
-
-            # Include gate type only when the node itself is a gate and it has
-            # at least one non-base child.  Previously only gate nodes were
-            # added to the model which meant that basic events—the actual
-            # causes in the chain—were omitted from the generated diagram. As a
-            # result the PDF report often displayed just the top event with no
-            # contributing causes.  By recording every node and linking it to
-            # its parent, all causes now appear in the output.
-            if node_type_up in GATE_NODE_TYPES and not self.all_children_are_base_events(node):
-                node_info["gate_type"] = node.gate_type
-            if getattr(node, "input_subtype", ""):
-                node_info["subtype"] = node.input_subtype
-
-            nodes.append(node_info)
-
-            for child in getattr(node, "children", []):
-                edges.append({"source": str(node.unique_id), "target": str(child.unique_id)})
-                traverse(child)
-
-        traverse(top_event)
-        return {"nodes": nodes, "edges": edges}
-
-    @staticmethod
-    def auto_generate_fta_diagram(fta_model, output_path):
-        """
-        Generate a cause-and-effect diagram with a layered (hierarchical) layout,
-        but draw the arrows in reverse (child -> parent).
-        """
-        import networkx as nx
-        from PIL import Image, ImageDraw, ImageFont
-        import numpy as np
-        import math
-        import sys
-
-        # --- 1) Build the directed graph (parent->child) ---
-        G = nx.DiGraph()
-        node_labels = {}
-        node_colors = {}
-
-        for node in fta_model["nodes"]:
-            node_id = node["id"]
-            label   = node.get("label", f"Node {node_id}")
-
-            # If there's a gate_type, append it to the label
-            gate_type = node.get("gate_type", "")
-            if gate_type:
-                # e.g. label = "Node 1\n(AND)"
-                label += f"\n({gate_type.upper()})"
-
-            G.add_node(node_id)
-            node_labels[node_id] = label
-
-            # Keep your color logic based on "subtype":
-            subtype = node.get("subtype", "").lower()
-            if "vehicle level function" in subtype:
-                node_colors[node_id] = "lightcoral"
-            elif "ai error" in subtype:
-                node_colors[node_id] = "lightyellow"
-            elif "failure" in subtype:
-                node_colors[node_id] = "lightblue"
-            elif "functional insufficiency" in subtype:
-                node_colors[node_id] = "lightgreen"
-            else:
-                node_colors[node_id] = "white"  # clone
-
-        # Add edges
-        for edge in fta_model["edges"]:
-            src = edge["source"]
-            tgt = edge["target"]
-            if not G.has_node(src) or not G.has_node(tgt):
-                continue
-            G.add_edge(src, tgt)
-
-        # --- 2) Identify the top event as 'root' (layer 0) ---
-        if fta_model["nodes"]:
-            top_event_id = fta_model["nodes"][0]["id"]
-        else:
-            img = Image.new("RGB", (400, 300), "white")
-            draw = ImageDraw.Draw(img)
-            draw.text((200, 150), "No nodes to display", fill="black", anchor="mm")
-            img.save(output_path)
-            return
-
-        # --- 3) BFS layering from top_event to find each node's layer ---
-        layers = {}
-        layers[top_event_id] = 0
-        queue = [top_event_id]
-        visited = set([top_event_id])
-
-        while queue:
-            current = queue.pop(0)
-            current_layer = layers[current]
-            for child in G.successors(current):
-                if child not in visited:
-                    visited.add(child)
-                    layers[child] = current_layer + 1
-                    queue.append(child)
-
-        # Any node not reached gets placed in a higher layer
-        max_layer = max(layers.values()) if layers else 0
-        for n in G.nodes():
-            if n not in layers:
-                max_layer += 1
-                layers[n] = max_layer
-
-        # Group nodes by layer
-        layer_dict = {}
-        for node_id, layer in layers.items():
-            layer_dict.setdefault(layer, []).append(node_id)
-
-        # --- 4) Assign (x, y) by layer ---
-        horizontal_gap = 2.0
-        vertical_gap   = 1.0
-        pos = {}
-
-        for layer in sorted(layer_dict.keys()):
-            node_list = layer_dict[layer]
-
-            # Sort siblings by average parent index (optional)
-            def avg_parent_position(n):
-                parents = list(G.predecessors(n))
-                if not parents:
-                    return 0
-                # we assume all parents are in a smaller layer
-                return sum(layer_dict[layers[p]].index(p) for p in parents) / len(parents)
-
-            node_list.sort(key=avg_parent_position)
-
-            # Place them at x = layer*gap, y around 0
-            middle = (len(node_list) - 1) / 2.0
-            for i, n in enumerate(node_list):
-                x = layer * horizontal_gap
-                y = (i - middle) * vertical_gap
-                pos[n] = (x, y)
-
-        # --- 5) Light collision-avoidance pass (optional) ---
-        def get_node_bbox(p, box_size=0.3):
-            return (p[0] - box_size, p[1] - box_size, p[0] + box_size, p[1] + box_size)
-
-        def bboxes_overlap(b1, b2):
-            return not (b1[2] < b2[0] or b1[0] > b2[2] or b1[3] < b2[1] or b1[1] > b2[3])
-
-        for _ in range(10):
-            for n1 in G.nodes():
-                for n2 in G.nodes():
-                    if n1 == n2:
-                        continue
-                    b1 = get_node_bbox(pos[n1])
-                    b2 = get_node_bbox(pos[n2])
-                    if bboxes_overlap(b1, b2):
-                        p1 = np.array(pos[n1])
-                        p2 = np.array(pos[n2])
-                        delta = p1 - p2
-                        dist = np.linalg.norm(delta) + 1e-9
-                        push = 0.02
-                        shift = (delta/dist)*push
-                        pos[n1] = tuple(p1 + shift)
-                        pos[n2] = tuple(p2 - shift)
-
-        # --- 6) Draw the diagram with REVERSED edges (child->parent) ---
-        # Convert layout coordinates to image pixels
-        xs = [p[0] for p in pos.values()]
-        ys = [p[1] for p in pos.values()]
-        min_x, max_x = min(xs), max(xs)
-        min_y, max_y = min(ys), max(ys)
-
-        # Ensure the canvas leaves enough room so that nodes at the
-        # extremities are fully visible.  The previous implementation used a
-        # fixed margin of 50 pixels which was smaller than half of the node's
-        # width (60px).  As a result, nodes located at the left or right
-        # boundary were clipped in the exported diagram.  By basing the margins
-        # on the node dimensions we guarantee that every node remains within
-        # view.
-        node_w, node_h = 120, 60
-        scale = 150
-        margin_x = int(node_w / 2) + 20
-        margin_y = int(node_h / 2) + 20
-        width = int((max_x - min_x) * scale) + 2 * margin_x
-        height = int((max_y - min_y) * scale) + 2 * margin_y
-
-        def to_px(pt):
-            x, y = pt
-            px = int((x - min_x) * scale) + margin_x
-            py = int((max_y - y) * scale) + margin_y
-            return px, py
-
-        px_pos = {n: to_px(pos[n]) for n in pos}
-
-        test_mod = sys.modules.get("test_cause_effect_diagram") or sys.modules.get("tests.test_cause_effect_diagram")
-        if test_mod and hasattr(test_mod, "created_sizes"):
-            test_mod.created_sizes.append((width, height))
-        img = Image.new("RGB", (width, height), "white")
-        draw = ImageDraw.Draw(img)
-        font = ImageFont.load_default()
-
-        # Draw reversed edges (child -> parent)
-        for src, tgt in G.edges():
-            start = px_pos[tgt]
-            end = px_pos[src]
-            draw.line([start, end], fill="gray", width=2)
-
-            dx = end[0] - start[0]
-            dy = end[1] - start[1]
-            length = math.hypot(dx, dy)
-            if length:
-                ux, uy = dx / length, dy / length
-                arrow = 10
-                left = (end[0] - ux * arrow - uy * arrow / 2,
-                        end[1] - uy * arrow + ux * arrow / 2)
-                right = (end[0] - ux * arrow + uy * arrow / 2,
-                         end[1] - uy * arrow - ux * arrow / 2)
-                draw.polygon([end, left, right], fill="gray")
-
-        # Draw nodes
-        for n, (x, y) in px_pos.items():
-            left = x - node_w / 2
-            top = y - node_h / 2
-            right = x + node_w / 2
-            bottom = y + node_h / 2
-            draw.rectangle(
-                [left, top, right, bottom],
-                fill=node_colors[n],
-                outline=StyleManager.get_instance().outline_color,
-            )
-            lbl = node_labels.get(n, str(n))
-            bbox = draw.multiline_textbbox((0, 0), lbl, font=font, align="center")
-            tw = bbox[2] - bbox[0]
-            th = bbox[3] - bbox[1]
-            draw.multiline_text((x - tw/2, y - th/2), lbl, font=font, fill="black", align="center")
-
-        img.save(output_path)
-
-    def build_dynamic_recommendations_table(events, app):
-        """
-        (Optional) If you still want to have a compact table of per-event recommendations,
-        this function builds a multiline LongTable with columns:
-        [Event Name, Prototype Assurance Level (PAL), Severity, Controllability, Description, Rationale, Dynamic Recommendations].
-        (Not used in the final report if you prefer only the consolidated argumentation.)
-        """
-        style_sheet = getSampleStyleSheet()
-        header_style = ParagraphStyle(
-            name="CompactHeader",
-            parent=style_sheet["BodyText"],
-            fontSize=4,
-            leading=5,
-            wordWrap='CJK',
-            alignment=1
-        )
-        body_style = ParagraphStyle(
-            name="CompactBody",
-            parent=style_sheet["BodyText"],
-            fontSize=5,
-            leading=6,
-            wordWrap='CJK'
-        )
-        
-        data = [[
-            Paragraph("<b>Event Name</b>", header_style),
-            Paragraph("<b>Prototype Assurance Level (PAL)</b>", header_style),
-            Paragraph("<b>Severity</b>", header_style),
-            Paragraph("<b>Controllability</b>", header_style),
-            Paragraph("<b>Description</b>", header_style),
-            Paragraph("<b>Rationale</b>", header_style),
-            Paragraph("<b>Recommendations</b>", header_style),
-        ]]
-        
-        for event in events:
-            event_name = event.name if event.name else f"Node {event.unique_id}"
-            if event.quant_value is not None:
-                disc_level = AutoML_Helper.discretize_level(event.quant_value)
-                assurance_str = app.assurance_level_text(disc_level)
-            else:
-                assurance_str = "N/A"
-            severity_str = str(event.severity) if event.severity is not None else "N/A"
-            controllability_str = str(event.controllability) if event.controllability is not None else "N/A"
-            desc_text = (event.description or "N/A").strip().replace("\n", "<br/>")
-            rat_text = (event.rationale or "N/A").strip().replace("\n", "<br/>")
-            rec_text = app.generate_argumentation_report(event)
-            if isinstance(rec_text, list):
-                rec_text = "\n".join(str(x) for x in rec_text)
-            rec_text = rec_text.strip().replace("\n", "<br/>")
-            data.append([
-                Paragraph(event_name, body_style),
-                Paragraph(assurance_str, body_style),
-                Paragraph(severity_str, body_style),
-                Paragraph(controllability_str, body_style),
-                Paragraph(desc_text, body_style),
-                Paragraph(rat_text, body_style),
-                Paragraph(rec_text, body_style),
-            ])
-        
-        col_widths = [100, 60, 40, 40, 150, 150, 200]
-        table = LongTable(data, colWidths=col_widths, repeatRows=1, splitByRow=1)
-        table.setStyle(TableStyle([
-            ('BACKGROUND', (0,0), (-1,0), colors.orange),
-            ('TEXTCOLOR', (0,0), (-1,0), colors.white),
-            ('ROWHEIGHT', (0,0), (-1,0), 12),
-            ('GRID', (0,0), (-1,-1), 0.25, colors.grey),
-            ('VALIGN', (0,0), (-1,-1), 'TOP'),
-            ('FONTSIZE', (0,1), (-1,-1), 5),
-            ('LEFTPADDING', (0,0), (-1,-1), 1),
-            ('RIGHTPADDING', (0,0), (-1,-1), 1),
-            ('TOPPADDING', (0,0), (-1,-1), 1),
-            ('BOTTOMPADDING', (0,0), (-1,-1), 1),
-            ('SPLITTABLE', (0,0), (-1,-1), True),
-        ]))
-        return table
-
-    def get_all_nodes_no_filter(self,node):
-        nodes = [node]
-        for child in node.children:
-            nodes.extend(self.get_all_nodes_no_filter(child))
-        return nodes
-        
-    def derive_requirements_for_event(self, event):
-        req_set = set()
-        for node in self.get_all_nodes(event):
-            if hasattr(node, "safety_requirements"):
-                for req in node.safety_requirements:
-                    req_set.add(f"[{req['id']}] [{req['req_type']}] {req['text']}")
-        return req_set
-
-    def get_combined_safety_requirements(self,node):
-        """
-        Returns a list of safety requirement dicts for the given node.
-        If the node is a clone, it also combines the original node's safety_requirements.
-        """
-        req_list = []
-        # Always take the node's own requirements if they exist.
-        if hasattr(node, "safety_requirements") and node.safety_requirements:
-            req_list.extend(node.safety_requirements)
-        # If node is a clone, also add requirements from its original.
-        if not node.is_primary_instance and hasattr(node, "original") and node.original.safety_requirements:
-            req_list.extend(node.original.safety_requirements)
-        return req_list
-
-    def get_top_event(self, node):
-        """
-        Walk up the parent chain until a node whose node_type is 'TOP EVENT' is found.
-        If none is found, return the node itself.
-        """
-        current = node
-        while current.parents:
-            for parent in current.parents:
-                if parent.node_type.upper() == "TOP EVENT":
-                    print(f"DEBUG: Found TOP EVENT for node {node.unique_id}: {parent.name}")
-                    return parent
-            current = current.parents[0]
-        print(f"DEBUG: No TOP EVENT found for node {node.unique_id}; returning self")
-        return node
-
-    def aggregate_safety_requirements(self, node, all_nodes):
-        aggregated = set()
-        # Always add the node’s own safety requirements.
-        for req in node.get("safety_requirements", []):
-            aggregated.add(req["id"])
-        
-        # If this node is a clone, also add the original’s aggregated safety requirements.
-        if node.get("original_id"):
-            original = all_nodes.get(node["original_id"])
-            if original:
-                aggregated.update(self.aggregate_safety_requirements(original, all_nodes))
-        
-        # NEW: Also add safety requirements from the node’s immediate parents.
-        for parent in node.get("parents", []):
-            for req in parent.get("safety_requirements", []):
-                aggregated.add(req["id"])
-        
-        # Recurse into children.
-        for child in node.get("children", []):
-            aggregated.update(self.aggregate_safety_requirements(child, all_nodes))
-        
-        node["aggregated_safety_requirements"] = sorted(aggregated)
-        return aggregated
-
-    def generate_top_event_summary(self, top_event):
-        """
-        Generates a structured, easy-to-read summary for a top-level event.
-        
-        It recursively collects all base nodes (nodes with type "CONFIDENCE LEVEL" or "ROBUSTNESS SCORE")
-        from the event’s entire subtree (using originals for clones) and then constructs a multi-line summary
-        that includes:
-          - The top-level event name.
-          - The required Prototype Assurance Level (PAL) (with numeric score) and the severity rating.
-          - A bullet-point list of base nodes with their scores and rationales.
-        """
-        # Retrieve all nodes from the entire subtree (including originals for clones)
-        all_nodes = self.get_all_nodes_no_filter(top_event)
-        
-        # Filter base nodes (confidence or robustness)
-        base_nodes = [n for n in all_nodes if n.node_type.upper() in ["CONFIDENCE LEVEL", "ROBUSTNESS SCORE"]]
-        
-        # Build a bullet list for base nodes
-        bullet_lines = []
-        for bn in base_nodes:
-            # Use the original's details for clones
-            orig = bn if bn.is_primary_instance else bn.original
-            identifier = orig.name if orig.name else f"Node {orig.unique_id}"
-            score = f"{orig.quant_value:.2f}" if orig.quant_value is not None else "N/A"
-            rationale = orig.rationale.strip() if orig.rationale and orig.rationale.strip() != "" else "No rationale provided"
-            bullet_lines.append(f"• {identifier}: Score = {score}, Rationale: {rationale}")
-        base_summary = "\n".join(bullet_lines) if bullet_lines else "No base nodes available."
-        
-        # Map overall assurance value to a descriptive level
-        overall_assurance = top_event.quant_value if top_event.quant_value is not None else 1.0
-        if overall_assurance >= 4.5:
-            assurance_descr = "PAL5"
-        elif overall_assurance >= 3.5:
-            assurance_descr = "PAL4"
-        elif overall_assurance >= 2.5:
-            assurance_descr = "PAL3"
-        elif overall_assurance >= 1.5:
-            assurance_descr = "PAL2"
-        else:
-            assurance_descr = "PAL1"
-        
-        # Use the top event's severity and controllability (defaults if missing)
-        try:
-            overall_severity = float(top_event.severity) if top_event.severity is not None else 3.0
-        except Exception:
-            overall_severity = 3.0
-        try:
-            overall_cont = float(top_event.controllability) if top_event.controllability is not None else 3.0
-        except Exception:
-            overall_cont = 3.0
-        
-        # Build the structured summary sentence
-        summary_sentence = (
-            f"Top-Level Event: {top_event.name}\n\n"
-            f"Assurance Requirement:\n"
-            f"  - Required Prototype Assurance Level (PAL): {assurance_descr} (Score: {overall_assurance:.2f})\n"
-            f"  - Severity Rating: {overall_severity:.2f}\n"
-            f"  - Controllability: {overall_cont:.2f}\n\n"
-            f"Rationale:\n"
-            f"  Based on analysis of its base nodes, the following factors contributed to this level:\n"
-            f"{base_summary}"
-        )
-        return summary_sentence
-
-    def _generate_pdf_report(self):
-        """Generate a PDF report based on the configurable template."""
-
-        report_title = self.project_properties.get(
-            "pdf_report_name", "AutoML-Analyzer PDF Report"
-        )
-        path = filedialog.asksaveasfilename(
-            defaultextension=".pdf", filetypes=[("PDF files", "*.pdf")]
-        )
-        if not path:
-            return
-
-        template_path = filedialog.askopenfilename(
-            title="Select Report Template",
-            defaultextension=".json",
-            filetypes=[("JSON files", "*.json")],
-            initialdir=_REPORT_TEMPLATE_PATH.parent,
-            initialfile=_REPORT_TEMPLATE_PATH.name,
-        )
-        if not template_path:
-            return
-
-        try:
-            from reportlab.lib.pagesizes import letter, landscape
-            from reportlab.lib.units import inch
-            from reportlab.platypus import (
-                Paragraph,
-                Spacer,
-                SimpleDocTemplate,
-                Image as RLImage,
-                Table,
-                TableStyle,
-                PageBreak,
-            )
-            from reportlab.lib.styles import getSampleStyleSheet, ParagraphStyle
-            from reportlab.lib import colors
-            from io import BytesIO
-            from PIL import Image, ImageDraw
-        except Exception:
-            messagebox.showerror(
-                "Report",
-                "Reportlab and Pillow packages are required to generate PDF reports.",
-            )
-            return
-
-        try:
-            template = load_report_template(template_path)
-        except Exception as exc:
-            messagebox.showerror("Report", f"Failed to load report template\n{exc}")
-            return
-
-        doc = SimpleDocTemplate(
-            path,
-            pagesize=landscape(letter),
-            leftMargin=0.8 * inch,
-            rightMargin=0.8 * inch,
-            topMargin=0.5 * inch,
-            bottomMargin=0.5 * inch,
-        )
-        styles = getSampleStyleSheet()
-        pdf_styles = styles
-        preformatted_style = ParagraphStyle(
-            name="Preformatted",
-            parent=pdf_styles["Normal"],
-            fontName="Courier",
-            fontSize=8,
-            leading=10,
-        )
-
-        def scale_image(img, max_width=500, max_height=300):
-            w, h = img.size
-            scale = min(max_width / w, max_height / h, 1)
-            return w * scale, h * scale
-
-        story = [Paragraph(report_title, styles["Title"]), Spacer(1, 12)]
-
-        def _element_base_matrix():
-            header_style = ParagraphStyle(
-                name="SafetyGoalsHeader",
-                parent=pdf_styles["Normal"],
-                fontSize=10,
-                leading=12,
-                alignment=1,
-            )
-            data = [
-                [
-                    Paragraph("<b>Robustness \\ Confidence</b>", header_style),
-                    Paragraph("<b>1 (Level 1)</b>", header_style),
-                    Paragraph("<b>2 (Level 2)</b>", header_style),
-                    Paragraph("<b>3 (Level 3)</b>", header_style),
-                    Paragraph("<b>4 (Level 4)</b>", header_style),
-                    Paragraph("<b>5 (Level 5)</b>", header_style),
-                ],
-                [
-                    Paragraph("<b>1 (Level 1)</b>", header_style),
-                    Paragraph("PAL5", pdf_styles["Normal"]),
-                    Paragraph("PAL5", pdf_styles["Normal"]),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                ],
-                [
-                    Paragraph("<b>2 (Level 2)</b>", header_style),
-                    Paragraph("PAL5", pdf_styles["Normal"]),
-                    Paragraph("PAL5", pdf_styles["Normal"]),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                ],
-                [
-                    Paragraph("<b>3 (Level 3)</b>", header_style),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                    Paragraph("PAL1", pdf_styles["Normal"]),
-                ],
-                [
-                    Paragraph("<b>4 (Level 4)</b>", header_style),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                    Paragraph("PAL1", pdf_styles["Normal"]),
-                    Paragraph("PAL1", pdf_styles["Normal"]),
-                ],
-                [
-                    Paragraph("<b>5 (Level 5)</b>", header_style),
-                    Paragraph("PAL4", pdf_styles["Normal"]),
-                    Paragraph("PAL3", pdf_styles["Normal"]),
-                    Paragraph("PAL1", pdf_styles["Normal"]),
-                    Paragraph("PAL1", pdf_styles["Normal"]),
-                    Paragraph("PAL1", pdf_styles["Normal"]),
-                ],
-            ]
-            table = Table(data, colWidths=[80, 70, 70, 70, 70, 70])
-            table.setStyle(
-                TableStyle(
-                    [
-                        ("BACKGROUND", (0, 0), (-1, 0), colors.lightblue),
-                        ("BACKGROUND", (0, 0), (0, -1), colors.lightblue),
-                        ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                        ("VALIGN", (0, 0), (-1, -1), "MIDDLE"),
-                        ("ALIGN", (0, 0), (-1, -1), "CENTER"),
-                        ("FONTSIZE", (0, 0), (-1, -1), 8),
-                    ]
-                )
-            )
-            return [
-                Paragraph("Table 1: Base Assurance Inversion Matrix", pdf_styles["Heading3"]),
-                Spacer(1, 6),
-                table,
-                Spacer(1, 12),
-            ]
-
-        def _element_discretization():
-            header_style = ParagraphStyle(
-                name="DiscretizationHeader",
-                parent=pdf_styles["Normal"],
-                fontSize=10,
-                leading=12,
-                alignment=1,
-            )
-            data = [
-                [
-                    Paragraph("<b>Continuous Value (Rounded)</b>", header_style),
-                    Paragraph("<b>Prototype Assurance Level (PAL)</b>", header_style),
-                ],
-                [Paragraph("< 1.5", header_style), Paragraph("Level 1 (PAL1)", pdf_styles["Normal"])],
-                [Paragraph("1.5 – < 2.5", header_style), Paragraph("Level 2 (PAL2)", pdf_styles["Normal"])],
-                [Paragraph("2.5 – < 3.5", header_style), Paragraph("Level 3 (PAL3)", pdf_styles["Normal"])],
-                [Paragraph("3.5 – < 4.5", header_style), Paragraph("Level 4 (PAL4)", pdf_styles["Normal"])],
-                [Paragraph("≥ 4.5", header_style), Paragraph("Level 5 (PAL5)", pdf_styles["Normal"])],
-            ]
-            table = Table(data, colWidths=[150, 200])
-            table.setStyle(
-                TableStyle(
-                    [
-                        ("BACKGROUND", (0, 0), (-1, 0), colors.lightblue),
-                        ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                        ("VALIGN", (0, 0), (-1, -1), "MIDDLE"),
-                        ("ALIGN", (0, 0), (-1, -1), "CENTER"),
-                        ("FONTSIZE", (0, 0), (-1, -1), 8),
-                    ]
-                )
-            )
-            return [
-                Paragraph("Table 2: Output Discretization Mapping", pdf_styles["Heading3"]),
-                Spacer(1, 6),
-                table,
-                Spacer(1, 12),
-            ]
-
-        def _element_hazop():
-            items: list = []
-            if self.hazop_docs:
-                items.append(PageBreak())
-                items.append(Paragraph("HAZOP Analyses", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for hz_doc in self.hazop_docs:
-                    items.append(Paragraph(hz_doc.name, pdf_styles["Heading3"]))
-                    data = [["Function", "Malfunction", "Hazard", "Safety"]]
-                    for e in hz_doc.entries:
-                        data.append([e.function, e.malfunction, e.hazard, "Yes" if e.safety else "No"])
-                    table = Table(data, repeatRows=1)
-                    table.setStyle(
-                        TableStyle(
-                            [
-                                ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                ("FONTSIZE", (0, 0), (-1, -1), 8),
-                            ]
-                        )
-                    )
-                    items.append(table)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_fi2tc():
-            items: list = []
-            if self.fi2tc_docs:
-                items.append(PageBreak())
-                items.append(Paragraph("FI2TC Analyses", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for fi_doc in self.fi2tc_docs:
-                    items.append(Paragraph(fi_doc.name, pdf_styles["Heading3"]))
-                    data = [["System Function", "Functional Insufficiencies", "Triggering Conditions", "Severity"]]
-                    for row in fi_doc.entries:
-                        data.append([
-                            row.get("system_function", ""),
-                            row.get("functional_insufficiencies", ""),
-                            row.get("triggering_conditions", ""),
-                            row.get("severity", ""),
-                        ])
-                    table = Table(data, repeatRows=1)
-                    table.setStyle(
-                        TableStyle(
-                            [
-                                ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                ("FONTSIZE", (0, 0), (-1, -1), 8),
-                            ]
-                        )
-                    )
-                    items.append(table)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_tc2fi():
-            items: list = []
-            if self.tc2fi_docs:
-                items.append(PageBreak())
-                items.append(Paragraph("TC2FI Analyses", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for tc_doc in self.tc2fi_docs:
-                    items.append(Paragraph(tc_doc.name, pdf_styles["Heading3"]))
-                    data = [["Known Use Case", "Functional Insufficiencies", "Triggering Conditions", "Severity"]]
-                    for row in tc_doc.entries:
-                        data.append([
-                            row.get("known_use_case", ""),
-                            row.get("functional_insufficiencies", ""),
-                            row.get("triggering_conditions", ""),
-                            row.get("severity", ""),
-                        ])
-                    table = Table(data, repeatRows=1)
-                    table.setStyle(
-                        TableStyle(
-                            [
-                                ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                ("FONTSIZE", (0, 0), (-1, -1), 8),
-                            ]
-                        )
-                    )
-                    items.append(table)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_risk():
-            items: list = []
-            if self.hara_docs:
-                items.append(PageBreak())
-                items.append(Paragraph("Risk Assessment", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for hara_doc in self.hara_docs:
-                    items.append(Paragraph(hara_doc.name, pdf_styles["Heading3"]))
-                    data = [[
-                        "Malfunction",
-                        "Hazard",
-                        "Severity",
-                        "Exposure",
-                        "Controllability",
-                        "ASIL",
-                        "Safety Goal",
-                    ]]
-                    for e in hara_doc.entries:
-                        data.append([
-                            e.malfunction,
-                            e.hazard,
-                            str(e.severity),
-                            str(e.exposure),
-                            str(e.controllability),
-                            e.asil,
-                            e.safety_goal,
-                        ])
-                    table = Table(data, repeatRows=1)
-                    table.setStyle(
-                        TableStyle(
-                            [
-                                ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                ("FONTSIZE", (0, 0), (-1, -1), 8),
-                            ]
-                        )
-                    )
-                    items.append(table)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_cbn():
-            items: list = []
-            if getattr(self, "cbn_docs", []):
-                items.append(PageBreak())
-                items.append(Paragraph("Causal Bayesian Network Analyses", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for cbn_doc in self.cbn_docs:
-                    items.append(Paragraph(cbn_doc.name, pdf_styles["Heading3"]))
-                    img = self.capture_cbn_diagram(cbn_doc)
-                    if img is not None:
-                        buf = BytesIO()
-                        img.save(buf, format="PNG")
-                        buf.seek(0)
-                        desired_width, desired_height = scale_image(img)
-                        rl_img = RLImage(buf, width=desired_width, height=desired_height)
-                        items.append(rl_img)
-                        items.append(Spacer(1, 12))
-                    network = cbn_doc.network
-                    for var in network.nodes:
-                        items.append(Paragraph(var, pdf_styles["Heading4"]))
-                        data = [["Combination", "P(True)", "P(Parents)", "P(All)"]]
-                        parents = network.parents.get(var, [])
-                        for combo, p_true, combo_prob, joint_prob in network.cpd_rows(var):
-                            combo_str = (
-                                ", ".join(f"{p}={v}" for p, v in zip(parents, combo))
-                                if parents
-                                else "(prior)"
-                            )
-                            data.append([
-                                combo_str,
-                                f"{p_true:.3f}",
-                                f"{combo_prob:.3f}",
-                                f"{joint_prob:.3f}",
-                            ])
-                        table = Table(data, repeatRows=1)
-                        table.setStyle(
-                            TableStyle(
-                                [
-                                    ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                    ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                    ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                    ("FONTSIZE", (0, 0), (-1, -1), 8),
-                                ]
-                            )
-                        )
-                        items.append(table)
-                        items.append(Spacer(1, 12))
-            return items
-
-        def _element_safety_goals():
-            items: list = []
-            level_labels = {1: "PAL1", 2: "PAL2", 3: "PAL3", 4: "PAL4", 5: "PAL5"}
-
-            def get_immediate_parent_assurance(node):
-                if node.parents:
-                    assurances = []
-                    for p in node.parents:
-                        parent = p if p.is_primary_instance else p.original
-                        try:
-                            val = int(parent.quant_value)
-                        except (TypeError, ValueError):
-                            val = 1
-                        assurances.append(val)
-                    return max(assurances) if assurances else int(
-                        node.quant_value if node.quant_value is not None else 1
-                    )
-                else:
-                    return int(node.quant_value if node.quant_value is not None else 1)
-
-            grouped_by_linked: dict = {}
-            for node in self.get_all_nodes_in_model():
-                if hasattr(node, "safety_requirements") and node.safety_requirements:
-                    safety_goal = (
-                        node.safety_goal_description.strip()
-                        if node.safety_goal_description.strip() != ""
-                        else node.name
-                    )
-                    parent_assur = get_immediate_parent_assurance(node)
-                    assurance_str = f"Level {parent_assur} ({level_labels.get(parent_assur, 'N/A')})"
-                    linked_rec = self.generate_recommendations_for_top_event(node)
-                    extra_recs = self.get_extra_recommendations_list(
-                        node.description, AutoML_Helper.discretize_level(node.quant_value)
-                    )
-                    if not extra_recs:
-                        extra_recs = ["No Extra Recommendation"]
-                    grouped_by_linked.setdefault(linked_rec, {})
-                    for extra in extra_recs:
-                        grouped_by_linked.setdefault(linked_rec, {}).setdefault(extra, [])
-                        grouped_by_linked[linked_rec][extra].append(
-                            f"- {safety_goal} (Assurance: {assurance_str})"
-                        )
-
-            sg_data = [
-                [
-                    Paragraph("<b>Linked Recommendation</b>", pdf_styles["Normal"]),
-                    Paragraph(
-                        "<b>Safety Goals Grouped by Extra Recommendation</b>",
-                        pdf_styles["Normal"],
-                    ),
-                ]
-            ]
-            for linked_rec, extra_groups in grouped_by_linked.items():
-                nested_text = ""
-                for extra_rec, goals in extra_groups.items():
-                    nested_text += f"<b>{extra_rec}:</b><br/>" + "<br/>".join(goals) + "<br/><br/>"
-                sg_data.append([
-                    Paragraph(linked_rec, pdf_styles["Normal"]),
-                    Paragraph(nested_text, pdf_styles["Normal"]),
-                ])
-            if len(sg_data) > 1:
-                table = Table(sg_data, colWidths=[200, 400])
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 10),
-                            ("ALIGN", (0, 0), (-1, 0), "CENTER"),
-                        ]
-                    )
-                )
-                items.append(Paragraph("Safety Goals Summary:", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                items.append(table)
-                items.append(Spacer(1, 12))
-                items.append(PageBreak())
-            return items
-
-        def _element_top_events():
-            items: list = []
-            cause_effect_rows = self.build_cause_effect_data()
-            processed_ids = set()
-            for idx, event in enumerate(self.top_events, start=1):
-                if event.unique_id in processed_ids:
-                    continue
-                processed_ids.add(event.unique_id)
-                items.append(
-                    Paragraph(
-                        f"Top-Level Event #{idx}: {event.name}", pdf_styles["Heading2"]
-                    )
-                )
-                items.append(Spacer(1, 12))
-                argumentation_text = self.generate_argumentation_report(event)
-                if isinstance(argumentation_text, list):
-                    argumentation_text = "\n".join(str(x) for x in argumentation_text)
-                argumentation_text = argumentation_text.replace("\n", "<br/>")
-                items.append(Paragraph(argumentation_text, preformatted_style))
-                items.append(Spacer(1, 12))
-
-                event_img = self.capture_event_diagram(event)
-                if event_img is not None:
-                    buf = BytesIO()
-                    event_img.save(buf, format="PNG")
-                    buf.seek(0)
-                    desired_width, desired_height = scale_image(event_img)
-                    rl_img = RLImage(buf, width=desired_width, height=desired_height)
-                    items.append(Paragraph("Detailed Diagram (Subtree):", pdf_styles["Heading3"]))
-                    items.append(Spacer(1, 12))
-                    items.append(rl_img)
-                    items.append(Spacer(1, 12))
-
-                ce_row = next(
-                    (
-                        r
-                        for r in cause_effect_rows
-                        if r["malfunction"] == getattr(event, "malfunction", "")
-                    ),
-                    None,
-                )
-                if ce_row:
-                    ce_img = self.render_cause_effect_diagram(ce_row)
-                    if ce_img:
-                        buf = BytesIO()
-                        ce_img.save(buf, format="PNG")
-                        buf.seek(0)
-                        desired_width, desired_height = scale_image(ce_img)
-                        rl_img2 = RLImage(buf, width=desired_width, height=desired_height)
-                        items.append(
-                            Paragraph("Cause and Effect Diagram:", pdf_styles["Heading3"])
-                        )
-                        items.append(Spacer(1, 12))
-                        items.append(rl_img2)
-                        items.append(Spacer(1, 12))
-                items.append(PageBreak())
-            return items
-
-        def _element_page_diagrams():
-            items: list = []
-            unique_page_nodes = {}
-            for evt in self.top_events:
-                for pg in self.get_page_nodes(evt):
-                    if pg.is_primary_instance:
-                        unique_page_nodes[pg.unique_id] = pg
-
-            if unique_page_nodes:
-                items.append(Paragraph("Page Diagrams:", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-
-            for page_node in unique_page_nodes.values():
-                page_img = self.capture_page_diagram(page_node)
-                if page_img is not None:
-                    buf = BytesIO()
-                    page_img.save(buf, format="PNG")
-                    buf.seek(0)
-                    desired_width, desired_height = scale_image(page_img)
-                    rl_page_img = RLImage(buf, width=desired_width, height=desired_height)
-                    items.append(
-                        Paragraph(
-                            f"Page Diagram for: {page_node.name}", pdf_styles["Heading3"]
-                        )
-                    )
-                    items.append(Spacer(1, 12))
-                    items.append(rl_page_img)
-                    items.append(Spacer(1, 12))
-                else:
-                    items.append(
-                        Paragraph(
-                            "A page diagram could not be captured.",
-                            pdf_styles["Normal"],
-                        )
-                    )
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_sysml_diagrams():
-            items: list = []
-            from mainappsrc.models.sysml.sysml_repository import SysMLRepository
-
-            repo = SysMLRepository.get_instance()
-            diagrams = list(repo.diagrams.values())
-            if diagrams:
-                items.append(Paragraph("SysML Diagrams:", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-
-            for diag in diagrams:
-                img = self.capture_sysml_diagram(diag)
-                if img is not None:
-                    buf = BytesIO()
-                    img.save(buf, format="PNG")
-                    buf.seek(0)
-                    desired_width, desired_height = scale_image(img)
-                    rl_img = RLImage(buf, width=desired_width, height=desired_height)
-                    items.append(Paragraph(diag.name, pdf_styles["Heading3"]))
-                    items.append(Spacer(1, 12))
-                    items.append(rl_img)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_fmea_tables():
-            items: list = []
-            if self.fmeas:
-                items.append(PageBreak())
-                items.append(Paragraph("FMEA Tables", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for fmea in self.fmeas:
-                    items.append(Paragraph(fmea['name'], pdf_styles["Heading3"]))
-                    data = [[
-                        "Component",
-                        "Parent",
-                        "Failure Mode",
-                        "Failure Effect",
-                        "Cause",
-                        "S",
-                        "O",
-                        "D",
-                        "RPN",
-                        "Requirements",
-                        "Malfunction",
-                    ]]
-                    for be in fmea['entries']:
-                        src = self.get_failure_mode_node(be)
-                        comp = self.get_component_name_for_node(src) or "N/A"
-                        parent = src.parents[0] if src.parents else None
-                        parent_name = (
-                            parent.user_name
-                            if parent and getattr(parent, "node_type", "").upper() not in GATE_NODE_TYPES
-                            else ""
-                        )
-                        req_ids = "; ".join([r.get("id") for r in getattr(be, 'safety_requirements', [])])
-                        rpn = be.fmea_severity * be.fmea_occurrence * be.fmea_detection
-                        failure_mode = be.description or (be.user_name or f"BE {be.unique_id}")
-                        row = [
-                            comp,
-                            parent_name,
-                            failure_mode,
-                            be.fmea_effect,
-                            getattr(be, 'fmea_cause', ''),
-                            be.fmea_severity,
-                            be.fmea_occurrence,
-                            be.fmea_detection,
-                            rpn,
-                            req_ids,
-                            getattr(be, 'fmeda_malfunction', ''),
-                        ]
-                        data.append(row)
-                    table = Table(data, repeatRows=1)
-                    table.setStyle(
-                        TableStyle(
-                            [
-                                ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                ("FONTSIZE", (0, 0), (-1, -1), 8),
-                            ]
-                        )
-                    )
-                    items.append(table)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_fmeda_tables():
-            items: list = []
-            if self.fmedas:
-                items.append(PageBreak())
-                items.append(Paragraph("FMEDA Tables", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                for fmeda in self.fmedas:
-                    items.append(Paragraph(fmeda['name'], pdf_styles["Heading3"]))
-                    data = [[
-                        "Component",
-                        "Parent",
-                        "Failure Mode",
-                        "Malfunction",
-                        "Safety Goal",
-                        "Fault Type",
-                        "Fraction",
-                        "FIT",
-                        "DiagCov",
-                        "Mechanism",
-                    ]]
-                    for be in fmeda['entries']:
-                        src = self.get_failure_mode_node(be)
-                        comp = self.get_component_name_for_node(src) or "N/A"
-                        parent = src.parents[0] if src.parents else None
-                        parent_name = (
-                            parent.user_name
-                            if parent and getattr(parent, "node_type", "").upper() not in GATE_NODE_TYPES
-                            else ""
-                        )
-                        failure_mode = be.description or (be.user_name or f"BE {be.unique_id}")
-                        row = [
-                            comp,
-                            parent_name,
-                            failure_mode,
-                            getattr(be, 'fmeda_malfunction', ''),
-                            getattr(be, 'fmeda_safety_goal', ''),
-                            getattr(be, 'fmeda_fault_type', ''),
-                            f"{getattr(be, 'fmeda_fault_fraction', 0)}",
-                            f"{getattr(be, 'fmeda_fit', 0)}",
-                            f"{getattr(be, 'fmeda_diag_cov', 0)}",
-                            getattr(be, 'fmeda_mechanism', ''),
-                        ]
-                        data.append(row)
-                    table = Table(data, repeatRows=1)
-                    table.setStyle(
-                        TableStyle(
-                            [
-                                ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                                ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                                ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                                ("FONTSIZE", (0, 0), (-1, -1), 8),
-                            ]
-                        )
-                    )
-                    items.append(table)
-                    items.append(Spacer(1, 12))
-            return items
-
-        def _element_traceability():
-            items: list = []
-            basic_events = [
-                n
-                for n in self.get_all_nodes(self.root_node)
-                if n.node_type.upper() == "BASIC EVENT"
-            ]
-            if basic_events:
-                items.append(PageBreak())
-                items.append(Paragraph("FTA-FMEA Traceability", pdf_styles["Heading2"]))
-                data = [["Basic Event", "Component"]]
-                for be in basic_events:
-                    comp = self.get_component_name_for_node(be) or "N/A"
-                    data.append([be.user_name or f"BE {be.unique_id}", comp])
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                items.append(table)
-                items.append(Spacer(1, 12))
-            return items
-
-        def _element_cut_sets():
-            items: list = []
-            cut_sets_exist = any(self.calculate_cut_sets(te) for te in self.top_events)
-            if cut_sets_exist:
-                items.append(PageBreak())
-                items.append(Paragraph("FTA Cut Sets", pdf_styles["Heading2"]))
-                data = [["Top Event", "Cut Set #", "Basic Events"]]
-                for te in self.top_events:
-                    nodes_by_id = {}
-
-                    def map_nodes(n):
-                        nodes_by_id[n.unique_id] = n
-                        for child in n.children:
-                            map_nodes(child)
-
-                    map_nodes(te)
-                    cut_sets = self.calculate_cut_sets(te)
-                    te_label = te.user_name or f"Top Event {te.unique_id}"
-                    for idx, cs in enumerate(cut_sets, start=1):
-                        names = ", ".join(
-                            f"{nodes_by_id[uid].user_name or nodes_by_id[uid].node_type} [{uid}]"
-                            for uid in sorted(cs)
-                        )
-                        data.append([te_label if idx == 1 else "", str(idx), names])
-                        te_label = ""
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                items.append(table)
-                items.append(Spacer(1, 12))
-            return items
-
-        def _element_common_cause():
-            items: list = []
-            events_by_cause = {}
-            for fmea in self.fmeas:
-                for be in fmea['entries']:
-                    cause = be.description
-                    label = f"{fmea['name']}:{be.user_name or be.description or be.unique_id}"
-                    events_by_cause.setdefault(cause, set()).add(label)
-            for fmeda in self.fmedas:
-                for be in fmeda['entries']:
-                    cause = be.description
-                    label = f"{fmeda['name']}:{be.user_name or be.description or be.unique_id}"
-                    events_by_cause.setdefault(cause, set()).add(label)
-            for be in self.get_all_basic_events():
-                cause = be.description or ""
-                label = be.user_name or f"BE {be.unique_id}"
-                events_by_cause.setdefault(cause, set()).add(label)
-            cc_rows = [[cause, ", ".join(sorted(evts))] for cause, evts in events_by_cause.items() if len(evts) > 1]
-            if cc_rows:
-                items.append(PageBreak())
-                items.append(Paragraph("Common Cause Analysis", pdf_styles["Heading2"]))
-                data = [["Cause", "Events"]] + cc_rows
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                items.append(table)
-                items.append(Spacer(1, 12))
-            return items
-
-        def _element_safety_security_reports():
-            items: list = []
-            library = getattr(self, "safety_case_library", None)
-            if library and getattr(library, "cases", None):
-                items.append(PageBreak())
-                items.append(Paragraph("Safety & Security Reports", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                data = [["Name"]]
-                for case in library.cases:
-                    data.append([case.name])
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                items.append(table)
-                items.append(Spacer(1, 12))
-            return items
-
-        def _element_spi_table():
-            items: list = []
-            tree = getattr(self, "_spi_tree", None)
-            if tree:
-                columns = list(tree["columns"])
-                data = [list(columns)]
-                for iid in tree.get_children(""):
-                    data.append(list(tree.item(iid, "values")))
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                items.append(table)
-                items.append(Spacer(1, 12))
-            return items
-
-        requirement_element_map = {
-            "req_vehicle": ("vehicle", "Vehicle"),
-            "req_operational": ("operational", "Operational"),
-            "req_operational_safety": ("operational safety", "Operational Safety"),
-            "req_functional_safety": ("functional safety", "Functional Safety"),
-            "req_technical_safety": ("technical safety", "Technical Safety"),
-            "req_ai_safety": ("AI safety", "AI Safety"),
-            "req_functional_modification": ("functional modification", "Functional Modification"),
-            "req_cybersecurity": ("cybersecurity", "Cybersecurity"),
-            "req_production": ("production", "Production"),
-            "req_service": ("service", "Service"),
-            "req_field_monitoring": ("field monitoring", "Field Monitoring"),
-            "req_decommissioning": ("decommissioning", "Decommissioning"),
-            "req_product": ("product", "Product"),
-            "req_legal": ("legal", "Legal"),
-            "req_organizational": ("organizational", "Organizational"),
-            "req_spi": ("spi", "Spi"),
-        }
-
-        def _make_requirement_table(req_type: str, title: str):
-            items: list = []
-            reqs = [
-                r for r in global_requirements.values() if r.get("req_type") == req_type
-            ]
-            if reqs:
-                items.append(PageBreak())
-                items.append(Paragraph(f"{title} Requirements", pdf_styles["Heading2"]))
-                items.append(Spacer(1, 12))
-                data = [["ID", "Text", "ASIL", "CAL"]]
-                for r in sorted(reqs, key=lambda x: x.get("id", "")):
-                    data.append(
-                        [
-                            r.get("id", ""),
-                            r.get("text", ""),
-                            r.get("asil", ""),
-                            r.get("cal", ""),
-                        ]
-                    )
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                items.append(table)
-                items.append(Spacer(1, 12))
-            return items
-
-        def _build_element(name: str, kind: str | None):
-            if kind:
-                if kind == "diagram":
-                    label = name
-                elif kind.startswith("diagram:"):
-                    label = kind.split(":", 1)[1]
-                else:
-                    label = None
-                if label is not None:
-                    img = Image.new("RGB", (400, 200), "white")
-                    if hasattr(img, "save"):
-                        draw = ImageDraw.Draw(img)
-                        if hasattr(draw, "rectangle"):
-                            draw.rectangle([0, 0, 399, 199], outline="black")
-                        if hasattr(draw, "text"):
-                            draw.text((10, 10), f"{label} diagram", fill="black")
-                        buf = BytesIO()
-                        img.save(buf, format="PNG")
-                        buf.seek(0)
-                        return [RLImage(buf)]
-                    return [Paragraph(f"[{label} diagram]", pdf_styles["Normal"])]
-                if kind.startswith("analysis:"):
-                    analysis_type = kind.split(":", 1)[1]
-                    return [
-                        Paragraph(
-                            f"[{analysis_type} analysis from diagrams]",
-                            pdf_styles["Normal"],
-                        )
-                    ]
-            if kind == "base_matrix":
-                return _element_base_matrix()
-            if kind == "discretization":
-                return _element_discretization()
-            if kind == "hazop":
-                return _element_hazop()
-            if kind == "fi2tc":
-                return _element_fi2tc()
-            if kind == "tc2fi":
-                return _element_tc2fi()
-            if kind == "risk":
-                return _element_risk()
-            if kind == "cbn":
-                return _element_cbn()
-            if kind == "safety_goals":
-                return _element_safety_goals()
-            if kind == "top_events":
-                return _element_top_events()
-            if kind == "page_diagrams":
-                return _element_page_diagrams()
-            if kind == "sysml_diagrams":
-                return _element_sysml_diagrams()
-            if kind == "fmea_tables":
-                return _element_fmea_tables()
-            if kind == "fmeda_tables":
-                return _element_fmeda_tables()
-            if kind == "traceability":
-                return _element_traceability()
-            if kind == "cut_sets":
-                return _element_cut_sets()
-            if kind == "common_cause":
-                return _element_common_cause()
-            if kind == "safety_security_reports":
-                return _element_safety_security_reports()
-            if kind == "spi_table":
-                return _element_spi_table()
-            if kind == "activity_actions":
-                data = [["Action"], ["Start"], ["Stop"]]
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                return [table]
-            if kind == "req_matrix_alloc":
-                data = [["Requirement", "Allocation"], ["REQ-1", "Block A"], ["REQ-2", "Block B"]]
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                return [table]
-            if kind == "product_goals":
-                return [
-                    Paragraph("Goal 1", pdf_styles["Normal"]),
-                    Paragraph("Goal 2", pdf_styles["Normal"]),
-                ]
-            if kind == "fsc_info":
-                return [
-                    Paragraph(
-                        "[functional safety concept information]",
-                        pdf_styles["Normal"],
-                    )
-                ]
-            if kind == "trace_matrix_pg_fsr":
-                data = [["Product Goal", "FSR"], ["PG1", "FSR1"], ["PG2", "FSR2"]]
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                return [table]
-            if kind == "trace_matrix_fsc":
-                data = [["TSR", "FSR"], ["TSR1", "FSR1"], ["TSR2", "FSR2"]]
-                table = Table(data, repeatRows=1)
-                table.setStyle(
-                    TableStyle(
-                        [
-                            ("BACKGROUND", (0, 0), (-1, 0), colors.lightgrey),
-                            ("GRID", (0, 0), (-1, -1), 0.5, colors.grey),
-                            ("VALIGN", (0, 0), (-1, -1), "TOP"),
-                            ("FONTSIZE", (0, 0), (-1, -1), 8),
-                        ]
-                    )
-                )
-                return [table]
-            if kind == "item_description":
-                text = getattr(self, "item_definition", {}).get("description", "")
-                if not text:
-                    text = "[item description]"
-                return [Paragraph(text, pdf_styles["Normal"])]
-            if kind == "assumptions":
-                text = getattr(self, "item_definition", {}).get("assumptions", "")
-                if not text:
-                    text = "[assumptions]"
-                return [Paragraph(text, pdf_styles["Normal"])]
-            if name in requirement_element_map:
-                req_type, title = requirement_element_map[name]
-                return _make_requirement_table(req_type, title)
-            return [Paragraph(f"[{name}]", styles["Normal"])]
-
-        elements = template.get("elements", {})
-        import re as _re
-
-        def _tokenize(text: str):
-            replaced = text
-            for placeholder in elements:
-                replaced = replaced.replace(
-                    f"<{placeholder}>", f"[[[element:{placeholder}]]]"
-                )
-            replaced = _re.sub(
-                r'<img\s+src="([^"]+)"\s*/?>',
-                lambda m: f"[[[img:{m.group(1)}]]]",
-                replaced,
-            )
-            replaced = _re.sub(
-                r'<a\s+href="([^"]+)">([^<]*)</a>',
-                lambda m: f"[[[link:{m.group(1)}|{m.group(2)}]]]",
-                replaced,
-            )
-            return _re.split(r"(\[\[\[[^\]]+\]\]\])", replaced)
-
-        for sec in template.get("sections", []):
-            story.append(Paragraph(sec.get("title", ""), styles["Heading2"]))
-            tokens = _tokenize(sec.get("content", ""))
-            for tok in tokens:
-                if not tok:
-                    continue
-                if tok.startswith("[[[") and tok.endswith("]]]"):
-                    inner = tok[3:-3]
-                    if inner.startswith("element:"):
-                        name = inner.split(":", 1)[1]
-                        story.extend(_build_element(name, elements.get(name)))
-                    elif inner.startswith("img:"):
-                        src = inner.split(":", 1)[1]
-                        try:
-                            story.append(RLImage(src))
-                        except Exception:
-                            story.append(
-                                Paragraph(f"[Missing image: {src}]", styles["Normal"])
-                            )
-                    elif inner.startswith("link:"):
-                        href, text = inner.split(":", 1)[1].split("|", 1)
-                        story.append(
-                            Paragraph(
-                                f'<link href="{href}">{text}</link>', styles["Normal"]
-                            )
-                        )
-                else:
-                    text = tok.strip()
-                    if text:
-                        for line in text.split("\n"):
-                            if line:
-                                story.append(Paragraph(line, styles["Normal"]))
-            story.append(Spacer(1, 12))
-
-        try:
-            doc.build(story)
-            json_path = Path(path).with_suffix(".json")
-            json_path.write_text(json.dumps(template, indent=2))
-            messagebox.showinfo("PDF Report", "PDF report generated!")
-        except Exception as exc:
-            messagebox.showerror("Report", f"Failed to generate PDF: {exc}")
-
-    def generate_pdf_report(self):
-        self._generate_pdf_report()
-
-    def capture_event_diagram(self, event_node):
-        temp = tk.Toplevel(self.root)
-        temp.withdraw()
-        canvas = tk.Canvas(temp, bg=StyleManager.get_instance().canvas_bg, width=2000, height=2000)
-        canvas.pack()
-        self.draw_subtree_with_filter(canvas, event_node, self.get_all_nodes(event_node))
-        canvas.delete("grid")
-        canvas.update()
-        bbox = canvas.bbox("all")
-        if not bbox:
-            temp.destroy()
-            return None
-        x, y, x2, y2 = bbox
-        margin_left = 100
-        margin_top  = 30
-        new_x = x - margin_left
-        new_y = y - margin_top
-        new_width = (x2 - x) + 2 * margin_left
-        new_height = (y2 - y) + 2 * margin_top
-        ps = canvas.postscript(colormode="color", x=new_x, y=new_y, width=new_width, height=new_height)
-        from io import BytesIO
-        ps_bytes = BytesIO(ps.encode("utf-8"))
-        try:
-            img = Image.open(ps_bytes)
-            img.load(scale=3)
-        except Exception:
-            img = None
-        temp.destroy()
-        return img.convert("RGB") if img else None
-
-    def draw_subtree_with_filter(self, canvas, root_event, visible_nodes):
+        return img.convert("RGB") if img else None    def draw_subtree_with_filter(self, canvas, root_event, visible_nodes):
         self.draw_connections_subtree(canvas, root_event, set())
         for n in visible_nodes:
             self.draw_node_on_canvas_pdf(canvas, n)
@@ -17576,571 +13121,24 @@ class AutoMLApp:
 
     # --- Review Toolbox Methods ---
     def start_peer_review(self):
-        dialog = ParticipantDialog(self.root, joint=False)
-
-        if dialog.result:
-            moderators, parts = dialog.result
-            name = simpledialog.askstring("Review Name", "Enter unique review name:")
-            if not name:
-                return
-            description = askstring_fixed(
-                simpledialog,
-                "Description",
-                "Enter a short description:",
-            )
-            if description is None:
-                description = ""
-            if not moderators:
-                messagebox.showerror("Review", "Please specify a moderator")
-                return
-            if not parts:
-                messagebox.showerror("Review", "At least one reviewer required")
-                return
-            due_date = simpledialog.askstring("Due Date", "Enter due date (YYYY-MM-DD):")
-            if any(r.name == name for r in self.reviews):
-                messagebox.showerror("Review", "Name already exists")
-                return
-            scope = ReviewScopeDialog(self.root, self)
-            (
-                fta_ids,
-                fmea_names,
-                fmeda_names,
-                hazop_names,
-                hara_names,
-                stpa_names,
-                fi2tc_names,
-                tc2fi_names,
-            ) = scope.result if scope.result else ([], [], [], [], [], [], [], [])
-            review = ReviewData(
-                name=name,
-                description=description,
-                mode='peer',
-                moderators=moderators,
-                participants=parts,
-                comments=[],
-                fta_ids=fta_ids,
-                fmea_names=fmea_names,
-                fmeda_names=fmeda_names,
-                hazop_names=hazop_names,
-                hara_names=hara_names,
-                stpa_names=stpa_names,
-                fi2tc_names=fi2tc_names,
-                tc2fi_names=tc2fi_names,
-                due_date=due_date,
-            )
-            self.reviews.append(review)
-            self.review_data = review
-            self.current_user = moderators[0].name if moderators else parts[0].name
-            self.open_review_document(review)
-            self.send_review_email(review)
-            self.open_review_toolbox()
+        return self.review_manager.start_peer_review()
 
     def start_joint_review(self):
-        dialog = ParticipantDialog(self.root, joint=True)
-        if dialog.result:
-            moderators, participants = dialog.result
-            name = simpledialog.askstring("Review Name", "Enter unique review name:")
-            if not name:
-                return
-            description = askstring_fixed(
-                simpledialog,
-                "Description",
-                "Enter a short description:",
-            )
-            if description is None:
-                description = ""
-            if not moderators:
-                messagebox.showerror("Review", "Please specify a moderator")
-                return
-            if not any(p.role == 'reviewer' for p in participants):
-                messagebox.showerror("Review", "At least one reviewer required")
-                return
-            if not any(p.role == 'approver' for p in participants):
-                messagebox.showerror("Review", "At least one approver required")
-                return
-            due_date = simpledialog.askstring("Due Date", "Enter due date (YYYY-MM-DD):")
-            if any(r.name == name for r in self.reviews):
-                messagebox.showerror("Review", "Name already exists")
-                return
-            scope = ReviewScopeDialog(self.root, self)
-            (
-                fta_ids,
-                fmea_names,
-                fmeda_names,
-                hazop_names,
-                hara_names,
-                stpa_names,
-                fi2tc_names,
-                tc2fi_names,
-            ) = scope.result if scope.result else ([], [], [], [], [], [], [], [])
-
-            # Ensure each selected element has a completed peer review
-            def peer_completed(pred):
-                return any(
-                    r.mode == 'peer'
-                    and getattr(r, 'reviewed', False)
-                    and pred(r)
-                    for r in self.reviews
-                )
-
-            for tid in fta_ids:
-                if not peer_completed(lambda r: tid in r.fta_ids):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_fta in fmea_names:
-                if not peer_completed(lambda r: name_fta in r.fmea_names):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_fd in fmeda_names:
-                if not peer_completed(lambda r: name_fd in r.fmeda_names):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_hz in hazop_names:
-                if not peer_completed(lambda r: name_hz in getattr(r, 'hazop_names', [])):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_hara in hara_names:
-                if not peer_completed(lambda r: name_hara in getattr(r, 'hara_names', [])):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_stpa in stpa_names:
-                if not peer_completed(lambda r: name_stpa in getattr(r, 'stpa_names', [])):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_fi in fi2tc_names:
-                if not peer_completed(lambda r: name_fi in getattr(r, 'fi2tc_names', [])):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            for name_tc in tc2fi_names:
-                if not peer_completed(lambda r: name_tc in getattr(r, 'tc2fi_names', [])):
-                    messagebox.showerror(
-                        "Review",
-                        "Peer review must be reviewed before starting joint review",
-                    )
-                    return
-            review = ReviewData(
-                name=name,
-                description=description,
-                mode='joint',
-                moderators=moderators,
-                participants=participants,
-                comments=[],
-                fta_ids=fta_ids,
-                fmea_names=fmea_names,
-                fmeda_names=fmeda_names,
-                hazop_names=hazop_names,
-                hara_names=hara_names,
-                stpa_names=stpa_names,
-                fi2tc_names=fi2tc_names,
-                tc2fi_names=tc2fi_names,
-                due_date=due_date,
-            )
-            self.reviews.append(review)
-            self.review_data = review
-            self.current_user = moderators[0].name if moderators else participants[0].name
-            self.open_review_document(review)
-            self.send_review_email(review)
-            self.open_review_toolbox()
+        return self.review_manager.start_joint_review()
 
     def open_review_document(self, review):
-        if hasattr(self, "_review_doc_tab") and self._review_doc_tab.winfo_exists():
-            self.doc_nb.select(self._review_doc_tab)
-        else:
-            title = f"Review {review.name}"
-            self._review_doc_tab = self._new_tab(title)
-            self._review_doc_window = ReviewDocumentDialog(self._review_doc_tab, self, review)
-            self._review_doc_window.pack(fill=tk.BOTH, expand=True)
-        self.refresh_all()
+        return self.review_manager.open_review_document(review)
 
     def open_review_toolbox(self):
-        if not self.reviews:
-            messagebox.showwarning("Review", "No reviews defined")
-            return
-        if not self.review_data and self.reviews:
-            self.review_data = self.reviews[0]
-        self.update_hara_statuses()
-        self.update_fta_statuses()
-        self.update_requirement_statuses()
-        if hasattr(self, "_review_tab") and self._review_tab.winfo_exists():
-            self.doc_nb.select(self._review_tab)
-        else:
-            self._review_tab = self._new_tab("Review")
-            self.review_window = ReviewToolbox(self._review_tab, self)
-            self.review_window.pack(fill=tk.BOTH, expand=True)
-        self.refresh_all()
-        self.user_manager.set_current_user()
+        return self.review_manager.open_review_toolbox()
 
     def send_review_email(self, review):
-        """Send the review summary to all reviewers via configured SMTP."""
-        recipients = [p.email for p in review.participants if p.role == 'reviewer' and p.email]
-        if not recipients:
-            return
+        return self.review_manager.send_review_email(review)
 
-        # Determine the current user's email if available
-        current_email = next((p.email for p in review.participants
-                              if p.name == self.current_user), "")
+    def capture_diff_diagram(self, top_event):
+        return self.review_manager.capture_diff_diagram(top_event)
 
-        if not getattr(self, "email_config", None):
-            cfg = EmailConfigDialog(self.root, default_email=current_email).result
-            self.email_config = cfg
-
-        cfg = getattr(self, "email_config", None)
-        if not cfg:
-            return
-
-        subject = f"Review: {review.name}"
-        lines = [f"Review Name: {review.name}", f"Description: {review.description}", ""]
-        if review.fta_ids:
-            lines.append("FTAs:")
-            for tid in review.fta_ids:
-                te = next((t for t in self.top_events if t.unique_id == tid), None)
-                if te:
-                    lines.append(f" - {te.name}")
-            lines.append("")
-        if review.fmea_names:
-            lines.append("FMEAs:")
-            for name in review.fmea_names:
-                lines.append(f" - {name}")
-            lines.append("")
-        if getattr(review, 'hazop_names', []):
-            lines.append("HAZOPs:")
-            for name in review.hazop_names:
-                lines.append(f" - {name}")
-            lines.append("")
-        if getattr(review, 'hara_names', []):
-            lines.append("Risk Assessments:")
-            for name in review.hara_names:
-                lines.append(f" - {name}")
-            lines.append("")
-        content = "\n".join(lines)
-        msg = EmailMessage()
-        msg['Subject'] = subject
-        msg['From'] = cfg['email']
-        msg['To'] = ', '.join(recipients)
-        msg.set_content(content)
-
-        html_lines = ["<html><body>", "<pre>", html.escape(content), "</pre>"]
-        image_cids = []
-        images = []
-        for tid in review.fta_ids:
-            node = self.find_node_by_id_all(tid)
-            if not node:
-                continue
-            img = self.capture_diff_diagram(node)
-            if img is None:
-                continue
-            buf = BytesIO()
-            img.save(buf, format="PNG")
-            buf.seek(0)
-            cid = make_msgid()
-            label = node.user_name or node.name or f"id{tid}"
-            html_lines.append(f"<p><b>FTA: {html.escape(label)}</b><br>" +
-                             f"<img src=\"cid:{cid[1:-1]}\" alt=\"{html.escape(label)}\"></p>")
-            image_cids.append(cid)
-            images.append(buf.getvalue())
-        diff_html = self.build_requirement_diff_html(review)
-        if diff_html:
-            html_lines.append("<b>Requirements:</b><br>" + diff_html)
-        html_lines.append("</body></html>")
-        html_body = "\n".join(html_lines)
-        msg.add_alternative(html_body, subtype="html")
-        html_part = msg.get_payload()[1]
-        for cid, data in zip(image_cids, images):
-            html_part.add_related(data, "image", "png", cid=cid)
-
-        # Attach FMEA tables as CSV files (can be opened with Excel)
-        for name in review.fmea_names:
-            fmea = next((f for f in self.fmeas if f["name"] == name), None)
-            if not fmea:
-                continue
-            out = StringIO()
-            writer = csv.writer(out)
-            columns = [
-                "Component",
-                "Parent",
-                "Failure Mode",
-                "Failure Effect",
-                "Cause",
-                "S",
-                "O",
-                "D",
-                "RPN",
-                "Requirements",
-            ]
-            writer.writerow(columns)
-            for be in fmea["entries"]:
-                src = self.get_failure_mode_node(be)
-                comp = self.get_component_name_for_node(src) or "N/A"
-                parent = src.parents[0] if src.parents else None
-                parent_name = parent.user_name if parent and getattr(parent, "node_type", "").upper() not in GATE_NODE_TYPES else ""
-                req_ids = "; ".join(
-                    [f"{req['req_type']}:{req['text']}" for req in getattr(be, 'safety_requirements', [])]
-                )
-                rpn = be.fmea_severity * be.fmea_occurrence * be.fmea_detection
-                failure_mode = be.description or (be.user_name or f"BE {be.unique_id}")
-                row = [
-                    comp,
-                    parent_name,
-                    failure_mode,
-                    be.fmea_effect,
-                    getattr(be, "fmea_cause", ""),
-                    be.fmea_severity,
-                    be.fmea_occurrence,
-                    be.fmea_detection,
-                    rpn,
-                    req_ids,
-                ]
-                writer.writerow(row)
-            csv_bytes = out.getvalue().encode('utf-8')
-            out.close()
-            msg.add_attachment(
-                csv_bytes,
-                maintype="text",
-                subtype="csv",
-                filename=f"fmea_{name}.csv",
-            )
-        for name in getattr(review, 'fmeda_names', []):
-            fmeda = next((f for f in self.fmedas if f["name"] == name), None)
-            if not fmeda:
-                continue
-            out = StringIO()
-            writer = csv.writer(out)
-            columns = [
-                "Component","Parent","Failure Mode","Malfunction","Safety Goal","FaultType","Fraction","FIT","DiagCov","Mechanism"
-            ]
-            writer.writerow(columns)
-            for be in fmeda["entries"]:
-                src = self.get_failure_mode_node(be)
-                comp = self.get_component_name_for_node(src) or "N/A"
-                parent = src.parents[0] if src.parents else None
-                parent_name = parent.user_name if parent and getattr(parent, "node_type", "").upper() not in GATE_NODE_TYPES else ""
-                row = [
-                    comp,
-                    parent_name,
-                    be.description or (be.user_name or f"BE {be.unique_id}"),
-                    be.fmeda_malfunction,
-                    be.fmeda_safety_goal,
-                    be.fmeda_fault_type,
-                    f"{be.fmeda_fault_fraction:.2f}",
-                    f"{be.fmeda_fit:.2f}",
-                    f"{be.fmeda_diag_cov:.2f}",
-                    getattr(be, "fmeda_mechanism", ""),
-                ]
-                writer.writerow(row)
-            csv_bytes = out.getvalue().encode('utf-8')
-            out.close()
-            msg.add_attachment(
-                csv_bytes,
-                maintype="text",
-                subtype="csv",
-                filename=f"fmeda_{name}.csv",
-            )
-        for name in getattr(review, 'hazop_names', []):
-            doc = next((d for d in self.hazop_docs if d.name == name), None)
-            if not doc:
-                continue
-            out = StringIO()
-            writer = csv.writer(out)
-            columns = [
-                "Function",
-                "Malfunction",
-                "Type",
-                "Scenario",
-                "Conditions",
-                "Hazard",
-                "Safety",
-                "Rationale",
-                "Covered",
-                "Covered By",
-            ]
-            writer.writerow(columns)
-            for e in doc.entries:
-                writer.writerow([
-                    self.get_entry_field(e, "function"),
-                    self.get_entry_field(e, "malfunction"),
-                    self.get_entry_field(e, "mtype"),
-                    self.get_entry_field(e, "scenario"),
-                    self.get_entry_field(e, "conditions"),
-                    self.get_entry_field(e, "hazard"),
-                    "Yes" if self.get_entry_field(e, "safety", False) else "No",
-                    self.get_entry_field(e, "rationale"),
-                    "Yes" if self.get_entry_field(e, "covered", False) else "No",
-                    self.get_entry_field(e, "covered_by"),
-                ])
-            csv_bytes = out.getvalue().encode("utf-8")
-            out.close()
-            msg.add_attachment(
-                csv_bytes,
-                maintype="text",
-                subtype="csv",
-                filename=f"hazop_{name}.csv",
-            )
-        for name in getattr(review, 'hara_names', []):
-            doc = next((d for d in self.hara_docs if d.name == name), None)
-            if not doc:
-                continue
-            out = StringIO()
-            writer = csv.writer(out)
-            columns = [
-                "Malfunction",
-                "Severity",
-                "Severity Rationale",
-                "Controllability",
-                "Cont. Rationale",
-                "Exposure",
-                "Exp. Rationale",
-                "ASIL",
-                "Safety Goal",
-            ]
-            writer.writerow(columns)
-            for e in doc.entries:
-                writer.writerow([
-                    self.get_entry_field(e, "malfunction"),
-                    self.get_entry_field(e, "severity"),
-                    self.get_entry_field(e, "sev_rationale"),
-                    self.get_entry_field(e, "controllability"),
-                    self.get_entry_field(e, "cont_rationale"),
-                    self.get_entry_field(e, "exposure"),
-                    self.get_entry_field(e, "exp_rationale"),
-                    self.get_entry_field(e, "asil"),
-                    self.get_entry_field(e, "safety_goal"),
-                ])
-            csv_bytes = out.getvalue().encode("utf-8")
-            out.close()
-            msg.add_attachment(
-                csv_bytes,
-                maintype="text",
-                subtype="csv",
-                filename=f"hara_{name}.csv",
-            )
-        try:
-            port = cfg.get('port', 465)
-            if port == 465:
-                with smtplib.SMTP_SSL(cfg['server'], port) as server:
-                    server.login(cfg['email'], cfg['password'])
-                    server.send_message(msg)
-            else:
-                with smtplib.SMTP(cfg['server'], port) as server:
-                    server.starttls()
-                    server.login(cfg['email'], cfg['password'])
-                    server.send_message(msg)
-        except smtplib.SMTPAuthenticationError:
-            messagebox.showerror(
-                "Email",
-                "Login failed. If your account uses two-factor authentication, "
-                "create an app password and use that instead of your normal password."
-            )
-            self.email_config = None
-        except (socket.gaierror, ConnectionRefusedError, smtplib.SMTPConnectError) as e:
-            messagebox.showerror(
-                "Email",
-                "Failed to connect to the SMTP server. Check the address, port and internet connection."
-            )
-            self.email_config = None
-        except Exception as e:
-            messagebox.showerror("Email", f"Failed to send review email: {e}")
-
-
-    def review_is_closed(self):
-        if not self.review_data:
-            return False
-        if getattr(self.review_data, "closed", False):
-            return True
-        if self.review_data.due_date:
-            try:
-                due = datetime.datetime.strptime(self.review_data.due_date, "%Y-%m-%d").date()
-                if datetime.date.today() > due:
-                    return True
-            except ValueError:
-                pass
-        return False
-
-    def review_is_closed_for(self, review):
-        if not review:
-            return False
-        if getattr(review, "closed", False):
-            return True
-        if review.due_date:
-            try:
-                due = datetime.datetime.strptime(review.due_date, "%Y-%m-%d").date()
-                if datetime.date.today() > due:
-                    return True
-            except ValueError:
-                pass
-        return False
-
-    def get_requirements_for_review(self, review):
-        """Return a set of requirement IDs included in the given review."""
-        req_ids = set()
-        for tid in getattr(review, "fta_ids", []):
-            node = self.find_node_by_id_all(tid)
-            if not node:
-                continue
-            for n in self.get_all_nodes(node):
-                for r in getattr(n, "safety_requirements", []):
-                    req_ids.add(r.get("id"))
-        for name in getattr(review, "fmea_names", []):
-            fmea = next((f for f in self.fmeas if f["name"] == name), None)
-            if not fmea:
-                continue
-            for e in fmea.get("entries", []):
-                for r in e.get("safety_requirements", []):
-                    req_ids.add(r.get("id"))
-        return req_ids
-
-    def update_requirement_statuses(self):
-        status_order = {
-            "draft": 0,
-            "in review": 1,
-            "peer reviewed": 2,
-            "pending approval": 3,
-            "approved": 4,
-        }
-        for req in global_requirements.values():
-            req.setdefault("status", "draft")
-        for review in self.reviews:
-            ids = self.get_requirements_for_review(review)
-            closed = self.review_is_closed_for(review)
-            for rid in ids:
-                req = global_requirements.get(rid)
-                if not req:
-                    continue
-                if review.mode == "joint":
-                    if review.approved:
-                        status = "approved"
-                    elif closed:
-                        status = "pending approval"
-                    else:
-                        status = "in review"
-                else:  # peer review
-                    if getattr(review, "reviewed", False) or closed:
-                        status = "peer reviewed"
-                    else:
-                        status = "in review"
-                if status_order[status] > status_order.get(req.get("status", "draft"), 0):
-                    req["status"] = status
-
+    # --- End Review Toolbox Methods ---
 
     def compute_requirement_asil(self, req_id):
         """Return highest ASIL across all safety goals linked to the requirement."""
